% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_functions.R, R/ext_tools.R
\name{aggr.amrfinder}
\alias{aggr.amrfinder}
\title{Callback function for aggregating AMRFinderPlus data for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.amrfinder(bascetFile, cellID, bascet_instance)

aggr.amrfinder(bascetFile, cellID, bascet_instance)
}
\value{
TODO

TODO
}
\description{
Callback function for aggregating AMRFinderPlus data for each cell.
To be called from BascetAggregateMap

Callback function for aggregating ABRicate data for each cell.
To be called from BascetAggregateMap
}
