% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{aggr.bakta}
\alias{aggr.bakta}
\title{Callback function for aggregating Bakta annotations for each cell}
\usage{
aggr.bakta(bascetFile, cellID, bascetInstance)
}
\arguments{
\item{bascetFile}{An opened Bascet file}

\item{cellID}{Cell ID}

\item{bascetInstance}{A Bascet instance}
}
\value{
Bakta annotations for one cell
}
\description{
To be called from BascetAggregateMap
}
