% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_functions.R
\name{aggr.checkm}
\alias{aggr.checkm}
\title{Callback function for aggregating CheckM data for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.checkm(bascetFile, cellID, bascet_instance)
}
\value{
TODO
}
\description{
Callback function for aggregating CheckM data for each cell.
To be called from BascetAggregateMap
}
