% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{aggr.gecco}
\alias{aggr.gecco}
\title{Callback function for aggregating GECCO data for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.gecco(bascetFile, cellID, bascet_instance)
}
\value{
TODO
}
\description{
Callback function for aggregating GECCO data for each cell.
To be called from BascetAggregateMap
}
