% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{aggr.minhash}
\alias{aggr.minhash}
\title{Callback function for aggregating min-hashes for each cell
To be called from BascetAggregateMap}
\usage{
aggr.minhash(bascetFile, cellID)
}
\value{
Minhash data (minhash.txt) for each cell
}
\description{
Callback function for aggregating min-hashes for each cell
To be called from BascetAggregateMap
}
