% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_functions.R
\name{aggr.rawtext}
\alias{aggr.rawtext}
\title{Callback function for just getting raw file contents
To be called from BascetAggregateMap}
\usage{
aggr.rawtext(fname)
}
\arguments{
\item{fname}{Name of output filename to get for each cell}
}
\value{
QUAST data for each cell
}
\description{
Callback function for just getting raw file contents
To be called from BascetAggregateMap
}
