% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{assembleBascetCommand}
\alias{assembleBascetCommand}
\title{Prepare Bascet command given arguments}
\usage{
assembleBascetCommand(bascetInstance, params)
}
\arguments{
\item{bascetInstance}{bascetInstance}

\item{params}{List of parameters}
}
\value{
Bascet command to run as a string
}
\description{
Prepare Bascet command given arguments
}
