% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{AtrandiBarcodeStats}
\alias{AtrandiBarcodeStats}
\title{Given a Bascet, produce a matrix showing for each combinatorial barcode,
how many times it occurs across the cells. Presented as a 96-well plate matrix}
\usage{
AtrandiBarcodeStats(
  bascetRoot,
  inputName = "debarcoded",
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard (should be debarcoded reads)}

\item{bascetInstance}{A Bascet instance}
}
\value{
Matrix showing coverage of each barcode
}
\description{
This command assumes that cells are named as follows: well1_well2_well3_well4,
where e.g. well1 is in format G12
}
