% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BarnyardPlotMatrix}
\alias{BarnyardPlotMatrix}
\title{Produce a matrix of Barnyard plots, i.e., counts for one species vs another,
for all combinations of species.}
\usage{
BarnyardPlotMatrix(adata)
}
\arguments{
\item{adata}{A Seurat object with the DefaultAssay holding counts per species (or similar)}
}
\value{
a ggarranged set of ggplots
}
\description{
Produce a matrix of Barnyard plots, i.e., counts for one species vs another,
for all combinations of species.
}
