% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{bascet_check_overwrite_output}
\alias{bascet_check_overwrite_output}
\title{Detect metadata for raw input FASTQ files}
\usage{
bascet_check_overwrite_output(outputFiles, overwrite)
}
\arguments{
\item{outputFiles}{Files that we expect to exist}

\item{overwrite}{Files that we expect to exist}
}
\value{
TRUE if ok to proceed
}
\description{
_R1 -- common from illumina sequencer
SRR****_1.fastq.gz -- typical from SRA
}
\details{
TODO Would be convenient to handle multiple samples, as sample1/xxx; in this case,
should prepend the sample name to the barcodes.

issue: when shardifying, good to keep info about what to merge. this reduces the work plenty!
could keep a list of which shards belong for the next step
}
