% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\docType{data}
\name{bascet_instance.default}
\alias{bascet_instance.default}
\title{The default Bascet installation settings}
\format{
An object of class \code{BascetInstance} of length 1.
}
\usage{
bascet_instance.default
}
\value{
A Bascet instance
}
\description{
The default Bascet installation settings
}
\keyword{datasets}
