% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateAMRfinder}
\alias{BascetAggregateAMRfinder}
\title{Aggregate data from AMRfinder
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateAMRfinder(
  bascetRoot,
  inputName = "AMRfinder",
  includeCells = NULL,
  getColumn = "Element.symbol",
  ...
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{includeCells}{Character vector of cell names to include, or NULL for all cells}

\item{getColumn}{Column name from AMRfinder output to use as feature (string)}

\item{...}{Additional arguments passed to \code{\link{BascetAggregateMap}}}
}
\value{
Aggregated data
}
\description{
Aggregate data from AMRfinder
This is a thin wrapper around BascetAggregateMap
}
\seealso{
\code{\link{BascetAggregateMap}}
}
