% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateAriba}
\alias{BascetAggregateAriba}
\title{Aggregate data from Ariba
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateAriba(bascetRoot, inputName = "ariba", includeCells = NULL, ...)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{includeCells}{Character vector of cell names to include, or NULL for all cells}

\item{...}{Additional arguments passed to \code{\link{BascetAggregateMap}}}
}
\value{
Aggregated data
}
\description{
Aggregate data from Ariba
This is a thin wrapper around BascetAggregateMap
}
\seealso{
\code{\link{BascetAggregateMap}}
}
