% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateGECCO}
\alias{BascetAggregateGECCO}
\title{Aggregate data from GECCO
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateGECCO(
  bascetRoot,
  inputName = "gecco",
  includeCells = NULL,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
Aggregated data
}
\description{
Aggregate data from GECCO
This is a thin wrapper around BascetAggregateMap
}
