% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{BascetAggregateMap}
\alias{BascetAggregateMap}
\title{Aggregate data from previous Map call}
\usage{
BascetAggregateMap(
  bascetRoot,
  bascetName,
  aggrFunction,
  showProgress = TRUE,
  bascet_instance
)
}
\value{
TODO
}
\description{
todo: allow multi-cpu support? parallel library
todo note, this is effectively a pure-R map function. different name?
}
