% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateQUAST}
\alias{BascetAggregateQUAST}
\title{Aggregate data from QUAST
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateQUAST(
  bascetRoot,
  inputName = "quast",
  includeCells = NULL,
  verbose = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\description{
Aggregate data from QUAST
This is a thin wrapper around BascetAggregateMap
}
