% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetAlignToReference}
\alias{BascetAlignToReference}
\title{Align from FASTQ, generate sorted and indexed BAM file}
\usage{
BascetAlignToReference(
  bascetRoot,
  useReference,
  numThreads = NULL,
  totalMem = NULL,
  inputName = "filtered",
  outputNameBAMcell = "aligned_cell",
  outputNameBAMpos = "aligned_pos",
  overwrite = FALSE,
  aligner = c(NULL, "BWA", "bowtie2", "STAR"),
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{useReference}{Name of the BWA reference to use}

\item{numThreads}{Number of threads to use for each runner. Default is the maximum, taken from runner settings}

\item{totalMem}{Total memory to allocate}

\item{inputName}{Name of input shard}

\item{outputNameBAMcell}{Name of cell-sorted BAMs}

\item{outputNameBAMpos}{Name of pos-sorted BAMs (if generated)}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\description{
Align from FASTQ, generate sorted and indexed BAM file
}
