% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetAssemble}
\alias{BascetAssemble}
\title{Assemble the genomes}
\usage{
BascetAssemble(
  bascetRoot,
  inputName = "rawreads",
  outputName = "assembled",
  runner,
  bascet_instance = bascet_instance.default
)
}
\description{
TODO not yet implemented
}
