% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetBam2Fragments}
\alias{BascetBam2Fragments}
\title{Take aligned BAM file and produce Fragments.tsv.gz, compatible with Signac ATAC-seq style analysis}
\usage{
BascetBam2Fragments(
  bascetRoot,
  inputName = "aligned",
  outputName = "fragments",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job, producing a type of Fragments.tsv.gz
}
\description{
Take aligned BAM file and produce Fragments.tsv.gz, compatible with Signac ATAC-seq style analysis
}
