% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetCacheComputation}
\alias{BascetCacheComputation}
\title{A wrapper to cache a computation. Put your function in as an argument,
as R will only compute its value if needed. If the cache file exist,
it will not be run again}
\usage{
BascetCacheComputation(bascetRoot, fname, value)
}
\arguments{
\item{fname}{Name of the file to store the cache in. The extension .RDS is added automatically}
}
\value{
The value
}
\description{
A wrapper to cache a computation. Put your function in as an argument,
as R will only compute its value if needed. If the cache file exist,
it will not be run again
}
