% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetCellNames}
\alias{BascetCellNames}
\title{Get list of cells in a Bascet}
\usage{
BascetCellNames(bascetRoot, bascetName)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{bascetName}{Name of the bascet}
}
\value{
Vector of cell names as strings
}
\description{
Get list of cells in a Bascet
}
