% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetCellNames_withstreamer}
\alias{BascetCellNames_withstreamer}
\title{Get list of cells in a Bascet -- streamer provided}
\usage{
BascetCellNames_withstreamer(bascetRoot, bascetName, streamer)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{bascetName}{Name of the bascet}

\item{streamer}{Bascet streamer instance}
}
\value{
Vector of cell names as strings
}
\description{
Get list of cells in a Bascet -- streamer provided
}
