% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{BascetComputeCountSketch}
\alias{BascetComputeCountSketch}
\title{Compute count sketch for each cell.
This is a thin wrapper around BascetMapCell}
\usage{
BascetComputeCountSketch(
  bascetRoot,
  inputName = "filtered",
  outputName = "countsketch",
  overwrite = FALSE,
  maxReads = 1e+05,
  kmerSize = 31,
  sketch_size = 5000,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
Compute count sketch for each cell.
This is a thin wrapper around BascetMapCell
}
