% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetCountChrom}
\alias{BascetCountChrom}
\title{From aligned BAM file, compute counts per chromosome}
\usage{
BascetCountChrom(
  bascetRoot,
  inputName = "aligned",
  outputName = "chromcount",
  min_matching = 0,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\value{
A job, executing the counting
}
\description{
From aligned BAM file, compute counts per chromosome
}
