% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{BascetCountMatrixToAssay}
\alias{BascetCountMatrixToAssay}
\title{Convert a BascetCountMatrix to a Seurat Assay}
\usage{
BascetCountMatrixToAssay(mat)
}
\arguments{
\item{mat}{A BascetCountMatrix object}
}
\value{
A Seurat AssayObject
}
\description{
Convert a BascetCountMatrix to a Seurat Assay
}
