% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetFeaturise}
\alias{BascetFeaturise}
\title{Select kmers that appear useful for clustering}
\usage{
BascetFeaturise(
  bascetRoot,
  inputName = "kmc",
  outputName = "sumkmers",
  includeCells = NULL,
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
CHANGED: it just sums up kmers
}
