% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetFeaturiseKMC}
\alias{BascetFeaturiseKMC}
\title{Select kmers that appear useful for clustering}
\usage{
BascetFeaturiseKMC(
  bascetRoot,
  inputName = "kmc",
  outputName = "sumkmers",
  includeCells = NULL,
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{includeCells}{List of cells to process}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
CHANGED: it just sums up kmers
}
