% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetFilterAlignment}
\alias{BascetFilterAlignment}
\title{Filter an alignment (BAM-file).}
\usage{
BascetFilterAlignment(
  bascetRoot,
  numLocalThreads = 1,
  inputName,
  outputName,
  keepMapped = FALSE,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{numLocalThreads}{Number of threads to use for each runner (TODO: get default from runner settings?)}

\item{inputName}{Name of input shards (BAM-file format)}

\item{outputName}{Name of output shards (BAM-file format)}

\item{keepMapped}{Keep the mapped reads (TRUE) or unmapped (FALSE)}
}
\description{
This is typically used to either remove host DNA, or keep reads mapping to a known reference
}
\details{
If the BAM-file has paired reads then BOTH reads need to be mapped (flag 0x2); otherwise (flag 0x4)
}
