% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{BascetGatherCountSketch}
\alias{BascetGatherCountSketch}
\title{Gather all count sketches into a single count sketch matrix}
\usage{
BascetGatherCountSketch(
  bascetRoot,
  inputName = "countsketch",
  outputName = "countsketch_mat.csv",
  includeCells = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{includeCells}{List of cells to process}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}
}
\value{
A job
}
\description{
TODO binary file format
}
