% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetGetRawAtrandiWGS}
\alias{BascetGetRawAtrandiWGS}
\title{Generate BAM with barcodes from input raw FASTQ}
\usage{
BascetGetRawAtrandiWGS(
  bascetRoot,
  rawmeta,
  outputName = "debarcoded",
  outputNameIncomplete = "incomplete_reads",
  chemistry = "atrandi_wgs",
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
Generate BAM with barcodes from input raw FASTQ
}
