% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeBowtie2}
\alias{BascetIndexGenomeBowtie2}
\title{Index a genome using Bowtie2 such that it can be used for alignment}
\usage{
BascetIndexGenomeBowtie2(
  genomeFile,
  numThreads = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{genomeFile}{Name of FASTA file holding genome sequence}

\item{numThreads}{Number of threads to use. Default is the maximum, taken from runner settings}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing if files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\description{
Index a genome using Bowtie2 such that it can be used for alignment
}
