% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeMinimap2}
\alias{BascetIndexGenomeMinimap2}
\title{Index a genome using minimap2 such that it can be used for alignment}
\usage{
BascetIndexGenomeMinimap2(
  fastaFile,
  indexFile,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{fastaFile}{Name of FASTA file holding genome sequence}

\item{indexFile}{Path to write the resulting .mmi index}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Index a genome using minimap2 such that it can be used for alignment
}
