% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeSTAR}
\alias{BascetIndexGenomeSTAR}
\title{Index a genome using STAR such that it can be used for alignment}
\usage{
BascetIndexGenomeSTAR(
  fastaFile,
  gtfFile,
  outDir,
  numLocalThreads = 10,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{gtfFile}{GFF file holding genome annotation}

\item{outDir}{A directory in which to store the index. This directory will be created}

\item{numLocalThreads}{The number of threads to use for the STAR index. Default is 10. TODO could get from runner}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}

\item{genomeFile}{Name of FASTA file holding genome sequence}
}
\description{
Index a genome using STAR such that it can be used for alignment
}
