% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{BascetInstance}
\alias{BascetInstance}
\title{Create a new bascet instance.
For advanced users only}
\usage{
BascetInstance(bin, tempdir, prependCmd = "", containerMem = "0B")
}
\arguments{
\item{bin}{Name of the binary}

\item{tempdir}{Directory where to store temporary files}

\item{prependCmd}{Something to prepend to the command, to e.g. support container systems}

\item{containerMem}{Amount of memory used by the container itself}
}
\value{
A Bascet instance
}
\description{
Create a new bascet instance.
For advanced users only
}
