% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetListFilesForCell}
\alias{BascetListFilesForCell}
\title{Read one file from a Bascet}
\usage{
BascetListFilesForCell(
  bascetFile,
  cellID,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
This can be made faster by, e.g., once and for all reading the location of
all objects in the file
}
