% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{BascetMakeKrakenCountMatrix}
\alias{BascetMakeKrakenCountMatrix}
\title{Produce a count matrix of taxonomy IDs from KRAKEN output}
\usage{
BascetMakeKrakenCountMatrix(
  bascetRoot,
  numLocalThreads = 1,
  inputName = "kraken_out",
  outputName = "kraken",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{numLocalThreads}{description}

\item{inputName}{description}

\item{outputName}{description}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{Configuration for how to run the Bascet Rust API}

\item{useKrakenDB}{description}
}
\description{
TODO: should run kraken in the same run
}
