% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetMakeMinhashHistogram}
\alias{BascetMakeMinhashHistogram}
\title{Gather all minhashes into a single histogram file}
\usage{
BascetMakeMinhashHistogram(
  bascetRoot,
  inputName = "minhash",
  outputName = "minhash_hist.csv",
  includeCells = NULL,
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{includeCells}{List of cells to process}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}
}
\value{
A job
}
\description{
Gather all minhashes into a single histogram file
}
