% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetMapCellAMRfinder}
\alias{BascetMapCellAMRfinder}
\title{Run AMRfinder on contigs of all cells.
This is a thin wrapper around BascetMapCell}
\usage{
BascetMapCellAMRfinder(
  bascetRoot,
  inputName = "contigs",
  outputName = "AMRfinder",
  db,
  ...
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{db}{Path to database}

\item{...}{Additional arguments passed to \code{\link{BascetMapCell}}}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Run AMRfinder on contigs of all cells.
This is a thin wrapper around BascetMapCell
}
\seealso{
\code{\link{BascetMapCell}}
}
