% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetMapCellAriba}
\alias{BascetMapCellAriba}
\title{....
This is a thin wrapper around BascetMapCell}
\usage{
BascetMapCellAriba(
  bascetRoot,
  inputName = "filtered",
  outputName = "ariba",
  db,
  overwrite = FALSE,
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{db}{description}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
....
This is a thin wrapper around BascetMapCell
}
