% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetMapCellGECCO}
\alias{BascetMapCellGECCO}
\title{Run GECCO on contigs of all cells.
This is a thin wrapper around BascetMapCell}
\usage{
BascetMapCellGECCO(
  bascetRoot,
  inputName = "contigs",
  outputName = "gecco",
  ...
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{...}{Additional arguments passed to \code{\link{BascetMapCell}}}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Run GECCO on contigs of all cells.
This is a thin wrapper around BascetMapCell
}
\seealso{
\code{\link{BascetMapCell}}
}
