% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetMapTransform}
\alias{BascetMapTransform}
\title{Transform: subset, convert, merge, divide}
\usage{
BascetMapTransform(
  bascetRoot,
  inputName,
  outputName,
  num_divide = 1,
  num_merge = 1,
  out_format = "tirp.gz",
  includeCells = NULL,
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
Transform: subset, convert, merge, divide
}
