% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetQuery}
\alias{BascetQuery}
\title{Build count table from kmer table and selected kmers}
\usage{
BascetQuery(
  bascetRoot,
  inputName = "kmc",
  outputName = "kmer",
  useKMERs,
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
Build count table from kmer table and selected kmers
}
