% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetQueryFq}
\alias{BascetQueryFq}
\title{Build count table from FASTQ reads and a list of selected kmers}
\usage{
BascetQueryFq(
  bascetRoot,
  inputName = "filtered",
  outputName = "kmer_counts",
  useKMERs,
  maxReads = 1e+06,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{useKMERs}{List of KMERs to query}

\item{maxReads}{The maximum number of reads per cell to sample}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Build count table from FASTQ reads and a list of selected kmers
}
