% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetQueryKMC}
\alias{BascetQueryKMC}
\title{Build count table from kmer table and a list of selected kmers}
\usage{
BascetQueryKMC(
  bascetRoot,
  inputName = "kmc",
  outputName = "kmer",
  useKMERs,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{useKMERs}{description}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
Build count table from kmer table and a list of selected kmers
}
