% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetReadFile}
\alias{BascetReadFile}
\title{Read one file from a Bascet}
\usage{
BascetReadFile(
  bascetFile,
  cellID,
  filename,
  as = c("tempfile"),
  bascetInstance = GetDefaultBascetInstance(),
  verbose = FALSE
)
}
\arguments{
\item{bascetFile}{Bascet file instance}

\item{filename}{description}

\item{as}{description}
}
\value{
Name of a temporary file where the read content is stored
}
\description{
TODO This can be made faster by, e.g., once and for all reading the location of
all objects in the file
}
