% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetReadMinhashHistogram}
\alias{BascetReadMinhashHistogram}
\title{Read histogram of KMERs, the output of BascetMakeMinhashHistogram}
\usage{
BascetReadMinhashHistogram(bascetRoot, inputName = "minhash_hist.csv")
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}
}
\value{
KMER histogram
}
\description{
Read histogram of KMERs, the output of BascetMakeMinhashHistogram
}
