% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_genomes.R
\name{BascetReadNcbiGenomeMetadata}
\alias{BascetReadNcbiGenomeMetadata}
\title{Read cached NCBI genome assembly metadata}
\usage{
BascetReadNcbiGenomeMetadata(path)
}
\arguments{
\item{path}{Path to \code{assembly_summary.txt}.}
}
\value{
A data frame with normalized column names.
}
\description{
Read cached NCBI genome assembly metadata
}
