% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{BascetReadObs}
\alias{BascetReadObs}
\title{Read observation metadata from a Bascet HDF5/AnnData file}
\usage{
BascetReadObs(fname)
}
\arguments{
\item{fname}{Path to a Bascet HDF5/AnnData file}
}
\value{
A data.frame with rownames set to cell names
}
\description{
Read observation metadata from a Bascet HDF5/AnnData file
}
