% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{BascetReadObsFromH5}
\alias{BascetReadObsFromH5}
\title{Read observation metadata from Bascet HDF5/AnnData files}
\usage{
BascetReadObsFromH5(bascetRoot, inputName, verbose = FALSE)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input HDF5 shard}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
A data.frame with rownames set to cell names
}
\description{
Read observation metadata from Bascet HDF5/AnnData files
}
