% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetRunCellSNP}
\alias{BascetRunCellSNP}
\title{Align from FASTQ, generate sorted and indexed BAM file}
\usage{
BascetRunCellSNP(
  bascetRoot,
  inputName = "aligned",
  outputName = "cellsnp",
  numThreads = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{numThreads}{Number of threads to use for each runner. Default is the maximum, taken from runner settings}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\description{
TODO Should be managed by Bascet mapshard system, with automatic input conversion. unaligned file should be made temp and removed
}
