% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetRunFASTP}
\alias{BascetRunFASTP}
\title{Run FASTP for each cell. Input must be in FASTQ file format}
\usage{
BascetRunFASTP(
  bascetRoot,
  numLocalThreads = 1,
  inputName = "asfq",
  outputName = "fastp",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\arguments{
\item{useKrakenDB}{Path to KRAKEN2 database}
}
\description{
Run FASTP for each cell. Input must be in FASTQ file format
}
