% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{BascetRunKraken}
\alias{BascetRunKraken}
\title{Run Kraken on each cell. TODO: should produce a count matrix right away!}
\usage{
BascetRunKraken(
  bascetRoot,
  useKrakenDB = "/data/henlab/kraken/standard-8",
  numLocalThreads = 1,
  inputName = "asfq",
  outputName = "kraken_out",
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
Run Kraken on each cell. TODO: should produce a count matrix right away!
}
