% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{BascetRunKrakenMakeMatrix}
\alias{BascetRunKrakenMakeMatrix}
\title{Run KRAKEN2 and produce a count matrix of taxonomy IDs}
\usage{
BascetRunKrakenMakeMatrix(
  bascetRoot,
  useKrakenDB = "/data/henlab/kraken/standard-8",
  numLocalThreads = 1,
  inputName = "kraken_out",
  outputName = "kraken",
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{useKrakenDB}{description}

\item{numLocalThreads}{description}

\item{inputName}{description}

\item{outputName}{description}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}
}
\description{
TODO: should run kraken in the same run
}
