% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetShardify}
\alias{BascetShardify}
\title{Take debarcoded reads and split them into suitable numbers of shards}
\usage{
BascetShardify(
  bascetRoot,
  inputName = "debarcoded",
  includeCells = NULL,
  num_output_shards = 1,
  outputName = "filtered",
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
Take debarcoded reads and split them into suitable numbers of shards
}
