% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetToFastq}
\alias{BascetToFastq}
\title{Convert data to Bascet-FASTQ}
\usage{
BascetToFastq(
  bascetRoot,
  inputName,
  outputName,
  numLocalThreads = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{numLocalThreads}{Number of threads to use per job. Default is the number from the runner}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Convert data to Bascet-FASTQ
}
