% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R, R/job_local.R, R/job_slurm.R
\name{CancelJob}
\alias{CancelJob}
\alias{CancelJob,NoJob-method}
\alias{CancelJob,LocalJob-method}
\alias{CancelJob,SlurmJob-method}
\title{Cancel a job}
\usage{
CancelJob(job)

\S4method{CancelJob}{NoJob}(job)

\S4method{CancelJob}{LocalJob}(job)

\S4method{CancelJob}{SlurmJob}(job)
}
\arguments{
\item{job}{A job object.}
}
\value{
Backend-specific cancellation result, usually invisible.
}
\description{
Cancel a job
}
\section{Methods (by class)}{
\itemize{
\item \code{CancelJob(NoJob)}: No-op job method.

\item \code{CancelJob(LocalJob)}: Local job method.

\item \code{CancelJob(SlurmJob)}: SLURM job method.

}}
