% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{ChromToSpeciesCount}
\alias{ChromToSpeciesCount}
\title{Produce a count matrix on strain level. too specific?}
\usage{
ChromToSpeciesCount(adata, map_seq2strain)
}
\value{
TODO
}
\description{
Produce a count matrix on strain level. too specific?
}
