% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{colnames,BascetCountMatrix-method}
\alias{colnames,BascetCountMatrix-method}
\title{Column names of a BascetCountMatrix (feature names)}
\usage{
\S4method{colnames}{BascetCountMatrix}(x)
}
\arguments{
\item{x}{A BascetCountMatrix object}
}
\value{
Character vector of column names
}
\description{
Column names of a BascetCountMatrix (feature names)
}
