% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{colnames<-,BascetCountMatrix-method}
\alias{colnames<-,BascetCountMatrix-method}
\title{Set column names of a BascetCountMatrix (feature names)}
\usage{
\S4method{colnames}{BascetCountMatrix}(x) <- value
}
\arguments{
\item{x}{A BascetCountMatrix object}

\item{value}{Character vector of new column names}
}
\value{
The modified BascetCountMatrix
}
\description{
Set column names of a BascetCountMatrix (feature names)
}
