% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{colSums,BascetCountMatrix-method}
\alias{colSums,BascetCountMatrix-method}
\title{Column sums of a BascetCountMatrix (counts per feature)}
\usage{
\S4method{colSums}{BascetCountMatrix}(x)
}
\arguments{
\item{x}{A BascetCountMatrix object}
}
\value{
Named numeric vector of column sums
}
\description{
Column sums of a BascetCountMatrix (counts per feature)
}
