% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{CountDataFrameToSparseMatrix}
\alias{CountDataFrameToSparseMatrix}
\title{Count entries in long format data frame and return as a sparse matrix}
\usage{
CountDataFrameToSparseMatrix(dat, rowname, colname)
}
\arguments{
\item{dat}{A data.frame}

\item{rowname}{Column to use as matrix row name}

\item{colname}{Column to use as matrix column name}
}
\value{
A dgCMatrix sparse matrix
}
\description{
Count entries in long format data frame and return as a sparse matrix
}
