% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{CreateSeuratObject.BascetCountMatrix}
\alias{CreateSeuratObject.BascetCountMatrix}
\title{Create a Seurat object from a BascetCountMatrix}
\usage{
CreateSeuratObject.BascetCountMatrix(counts, ...)
}
\arguments{
\item{counts}{A BascetCountMatrix object}

\item{...}{Additional arguments passed to the default CreateSeuratObject method}
}
\value{
A Seurat object
}
\description{
Create a Seurat object from a BascetCountMatrix
}
