% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{CreateSeuratObjectWithReduction}
\alias{CreateSeuratObjectWithReduction}
\title{Create a seurat object from e.g. count sketch reduction}
\usage{
CreateSeuratObjectWithReduction(
  Q,
  reduction_name = "kmersketch",
  assay = "RNA"
)
}
\value{
Seurat object; make an assay?
}
\description{
Create a seurat object from e.g. count sketch reduction
}
