% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_slurm.R
\name{createSlurmJobFromExisting}
\alias{createSlurmJobFromExisting}
\title{This creates a job object, linking to a running command.
Mainly used for development but can be used in case
a Zorn session died and you want to create a new
monitor}
\usage{
createSlurmJobFromExisting(pid, arraysize)
}
\arguments{
\item{pid}{Process ID}

\item{arraysize}{Size of array (must be known)}
}
\value{
A SLURM job object
}
\description{
This creates a job object, linking to a running command.
Mainly used for development but can be used in case
a Zorn session died and you want to create a new
monitor
}
