% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoding.R
\name{DebarcodedKneePlot}
\alias{DebarcodedKneePlot}
\title{Produce summary kneeplot given debarcoded statistics}
\usage{
DebarcodedKneePlot(debstat, filename = NULL)
}
\arguments{
\item{debstat}{Statistics produced about debarcoded files}

\item{filename}{Optional file to store kneeplot in}
}
\value{
A ggplot object
}
\description{
Produce summary kneeplot given debarcoded statistics
}
