% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{defaultBascetBinDir}
\alias{defaultBascetBinDir}
\title{Get the default directory for storing Bascet binaries}
\usage{
defaultBascetBinDir()
}
\value{
A directory path
}
\description{
Get the default directory for storing Bascet binaries
}
