% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{detect_shards_for_file}
\alias{detect_shards_for_file}
\title{Helper function: Figure out which shards belong together given root input name and extension
i.e. root/name.##.ext}
\usage{
detect_shards_for_file(bascetRoot, inputName)
}
\description{
Helper function: Figure out which shards belong together given root input name and extension
i.e. root/name.##.ext
}
