% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{DetectRawFileMeta}
\alias{DetectRawFileMeta}
\title{Detect metadata for raw input FASTQ files}
\usage{
DetectRawFileMeta(rawRoot, verbose = FALSE)
}
\arguments{
\item{rawRoot}{Path to folder with FASTQ files}
}
\value{
A data frame with metadata for the raw input files
}
\description{
_R1 -- common from illumina sequencer
SRR****_1.fastq.gz -- typical from SRA
}
\details{
TODO Would be convenient to handle multiple samples, as sample1/xxx; in this case,
should prepend the sample name to the barcodes.

issue: when shardifying, good to keep info about what to merge. this reduces the work plenty!
could keep a list of which shards belong for the next step
}
