% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{detectShardsForFile}
\alias{detectShardsForFile}
\title{Helper function: Figure out which shards belong together given root input name and extension
i.e. root/name.##.ext}
\usage{
detectShardsForFile(bascetRoot, inputName)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}
}
\value{
List of files that correspond to the shard
}
\description{
Helper function: Figure out which shards belong together given root input name and extension
i.e. root/name.##.ext
}
