% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{dim,BascetCountMatrix-method}
\alias{dim,BascetCountMatrix-method}
\title{Dimensions of a BascetCountMatrix}
\usage{
\S4method{dim}{BascetCountMatrix}(x)
}
\arguments{
\item{x}{A BascetCountMatrix object}
}
\value{
Integer vector of length 2 (rows, columns) of the count matrix
}
\description{
Dimensions of a BascetCountMatrix
}
