% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseAMRfinder}
\alias{DownloadDatabaseAMRfinder}
\title{Download a database for AMRfinder}
\usage{
DownloadDatabaseAMRfinder(dbdir, bascetInstance = GetDefaultBascetInstance())
}
\arguments{
\item{dbdir}{Directory to download database to}

\item{bascetInstance}{A Bascet instance}
}
\description{
Download a database for AMRfinder
}
