% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseAriba}
\alias{DownloadDatabaseAriba}
\title{Download database for Ariba}
\usage{
DownloadDatabaseAriba(
  dbdir,
  ref = c("ncbi", "argannot", "card", "megares", "plasmidfinder", "resfinder",
    "srst2_argannot", "vfdb_core", "vfdb_full", "virulencefinder"),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{dbdir}{Directory in which to download database}

\item{ref}{Which reference to download}

\item{bascetInstance}{A Bascet instance}
}
\description{
Download database for Ariba
}
