% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseBakta}
\alias{DownloadDatabaseBakta}
\title{Download a database for Bakta}
\usage{
DownloadDatabaseBakta(
  dbdir,
  dbtype = c("light", "full"),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{dbdir}{Directory path to download database to (string)}

\item{dbtype}{Database type: "light" or "full"}

\item{bascetInstance}{A Bascet instance}
}
\description{
Download a database for Bakta
}
