% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_exit}
\alias{extractstreamer_exit}
\title{extract streamer: end an instance.
the object should no longer be used after calling this function}
\usage{
extractstreamer_exit(p)
}
\description{
extract streamer: end an instance.
the object should no longer be used after calling this function
}
