% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_extract_to}
\alias{extractstreamer_extract_to}
\title{extract streamer: extract to external file}
\usage{
extractstreamer_extract_to(p, fname, outpath)
}
\value{
0 if ok
}
\description{
extract streamer: extract to external file
}
