% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_ls}
\alias{extractstreamer_ls}
\title{extract streamer: list all files in current zip-file}
\usage{
extractstreamer_ls(p, super_verbose = FALSE)
}
\value{
list of files
}
\description{
extract streamer: list all files in current zip-file
}
