% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_open}
\alias{extractstreamer_open}
\title{extract streamer: set which file is open}
\usage{
extractstreamer_open(p, fname)
}
\value{
0 if ok
}
\description{
extract streamer: set which file is open
}
