% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_read_n_lines}
\alias{extractstreamer_read_n_lines}
\title{extract streamer: helper function to read N lines}
\usage{
extractstreamer_read_n_lines(p, n_lines, verbose = FALSE)
}
\value{
all the lines
}
\description{
extract streamer: helper function to read N lines
}
