% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_read_one_line}
\alias{extractstreamer_read_one_line}
\title{extract streamer: helper function to read one line}
\usage{
extractstreamer_read_one_line(p, super_verbose)
}
\value{
the line
}
\description{
extract streamer: helper function to read one line
}
