% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_showtext}
\alias{extractstreamer_showtext}
\title{extract streamer: get content of file, assumed to be text (or this function crashes)}
\usage{
extractstreamer_showtext(p, fname)
}
\value{
The text, divided by line
}
\description{
extract streamer: get content of file, assumed to be text (or this function crashes)
}
