% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamer_start}
\alias{extractstreamer_start}
\title{extract streamer: create an instance}
\usage{
extractstreamer_start(
  fname = NULL,
  verbose = FALSE,
  bascet_instance = GetDefaultBascetInstance()
)
}
\value{
The process
}
\description{
extract streamer: create an instance
}
