% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerExit}
\alias{extractstreamerExit}
\title{extract streamer: end an instance.
the object should no longer be used after calling this function}
\usage{
extractstreamerExit(p)
}
\arguments{
\item{p}{Streamer process}
}
\description{
extract streamer: end an instance.
the object should no longer be used after calling this function
}
