% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerExtractTo}
\alias{extractstreamerExtractTo}
\title{extract streamer: extract to external file}
\usage{
extractstreamerExtractTo(p, fname, outpath)
}
\value{
0 if ok
}
\description{
extract streamer: extract to external file
}
