% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerLs}
\alias{extractstreamerLs}
\title{extract streamer: list all files in current zip-file}
\usage{
extractstreamerLs(p, superVerbose = FALSE)
}
\value{
list of files
}
\description{
extract streamer: list all files in current zip-file
}
