% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerOpen}
\alias{extractstreamerOpen}
\title{extract streamer: set which file is open}
\usage{
extractstreamerOpen(p, fname, verbose = FALSE)
}
\arguments{
\item{p}{Streamer process}

\item{fname}{Name of file}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
0 if ok
}
\description{
extract streamer: set which file is open
}
