% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerReadOneLine}
\alias{extractstreamerReadOneLine}
\title{extract streamer: helper function to read one line}
\usage{
extractstreamerReadOneLine(p, superVerbose = FALSE)
}
\value{
the line
}
\description{
extract streamer: helper function to read one line
}
