% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerShowtext}
\alias{extractstreamerShowtext}
\title{extract streamer: get content of file, assumed to be text (or this function crashes)}
\usage{
extractstreamerShowtext(p, fname, verbose = FALSE)
}
\arguments{
\item{p}{Streamer process}

\item{fname}{Name of file}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
The text, divided by line
}
\description{
extract streamer: get content of file, assumed to be text (or this function crashes)
}
