% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerStart}
\alias{extractstreamerStart}
\title{extract streamer: create an instance}
\usage{
extractstreamerStart(
  fname = NULL,
  verbose = FALSE,
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{fname}{File to stream (can be NULL if set later)}

\item{verbose}{Print additional information, primarily to help troubleshooting}

\item{bascetInstance}{A Bascet instance}
}
\value{
The process
}
\description{
extract streamer: create an instance
}
