% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{format_size_bascet}
\alias{format_size_bascet}
\title{Format size for input to Bascet}
\usage{
format_size_bascet(s)
}
\arguments{
\item{s}{Size as output from fs::fs_bytes, or a string}
}
\description{
Format size for input to Bascet
}
