% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{FragmentCountsPerChrom}
\alias{FragmentCountsPerChrom}
\title{From a Signac chromatin assay with fragments, for each cell, count how many reads per chromosome}
\usage{
FragmentCountsPerChrom(chrom_assay)
}
\value{
TODO
}
\description{
From a Signac chromatin assay with fragments, for each cell, count how many reads per chromosome
}
