% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{FragmentCountsPerChromAssay}
\alias{FragmentCountsPerChromAssay}
\title{From a Signac chromatin assay with fragments, for each cell, count how many reads per chromosome.
This function directly returns an assay that can be added to a Seurat multimodal object}
\usage{
FragmentCountsPerChromAssay(bascetRoot, inputName = "fragments.1.tsv.gz")
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}
}
\value{
A seurat object holding the counts
}
\description{
From a Signac chromatin assay with fragments, for each cell, count how many reads per chromosome.
This function directly returns an assay that can be added to a Seurat multimodal object
}
