% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{FragmentsToSignac}
\alias{FragmentsToSignac}
\title{From a fragments file, get a chromatin assay for Signac.}
\usage{
FragmentsToSignac(fragpath)
}
\arguments{
\item{fragpath}{List to fragment file}
}
\value{
A ChromatinAssay
}
\description{
Note: signac is dirt slow at counting as of writing. It might be scalable enough
for certain inputs and tasks, but we still provide the option of using it.
}
