% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{get_fastq_R2_from_R1}
\alias{get_fastq_R2_from_R1}
\title{Get corresponding R2 file. Assumes that the input file is R1}
\usage{
get_fastq_R2_from_R1(fname)
}
\arguments{
\item{fname}{Path to file}
}
\value{
Path to R2 file
}
\description{
Get corresponding R2 file. Assumes that the input file is R1
}
