% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetBinaryDev}
\alias{getBascetBinaryDev}
\title{Get a Bascet binary for the current platform from the development server
It will be cached in the provided directory to avoid downloading it each the time the function is called}
\usage{
getBascetBinaryDev(
  storeAt = NULL,
  tempdir = NULL,
  logLevel = "info",
  forceInstall = FALSE,
  containerMem = "2GB"
)
}
\arguments{
\item{storeAt}{Directory to store the binary in. If NULL, uses \code{\link{defaultBascetBinDir}}}

\item{tempdir}{Default is to create a directory for temporary files in the current directory. Place it on a fast disk if possible}

\item{logLevel}{Log level for the Bascet instance (e.g. "info", "debug", "warn")}

\item{forceInstall}{Force download of the Bascet binary even if a cached binary exists}

\item{containerMem}{Amount of memory used by the container itself}
}
\value{
A Bascet instance
}
\description{
Get a Bascet binary for the current platform from the development server
It will be cached in the provided directory to avoid downloading it each the time the function is called
}
