% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetDevDir}
\alias{getBascetDevDir}
\title{Get a Bascet binary from the target from a locally built Bascet repository}
\usage{
getBascetDevDir(
  devdir,
  tempdir = NULL,
  logLevel = "info",
  targetType = "release"
)
}
\arguments{
\item{tempdir}{Default is to create a directory for temporary files in the current directory. Place it on a fast disk if possible}

\item{logLevel}{Log level for the Bascet instance (e.g. "info", "debug", "warn")}

\item{targetType}{What target type to load}

\item{storeAt}{Directory to store the container in. Default is current directory but it is likely better to provide a single systems level directory #################### TODO}
}
\value{
A Bascet instance
}
\description{
Get a Bascet binary from the target from a locally built Bascet repository
}
