% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetDockerImage}
\alias{getBascetDockerImage}
\title{Get and install a Bascet docker image.}
\usage{
getBascetDockerImage(
  storeAt = getwd(),
  tempdir = NULL,
  forceInstall = FALSE,
  verbose = FALSE
)
}
\value{
A Bascet instance
}
\description{
Get and install a Bascet docker image.
}
