% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetExecutable}
\alias{getBascetExecutable}
\title{Get a Bascet executable
It will be cached in the provided directory to avoid downloading it each the time the function is called}
\usage{
getBascetExecutable(storeAt = getwd(), tempdir = NULL, logLevel = "info")
}
\arguments{
\item{storeAt}{Directory to store the container in. Default is current directory but it is likely better to provide a single systems level directory #################### TODO}

\item{tempdir}{Default is to create a directory for temporary files in the current directory. Place it on a fast disk if possible}

\item{logLevel}{Log level for the Bascet instance (e.g. "info", "debug", "warn")}
}
\value{
A Bascet instance
}
\description{
Get a Bascet executable
It will be cached in the provided directory to avoid downloading it each the time the function is called
}
