% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetImageInstance}
\alias{getBascetImageInstance}
\title{Get a Bascet image (singularity or docker).
It will be cached in the provided directory to avoid downloading it all the time}
\usage{
getBascetImageInstance(store_at = "./", tempdir = NULL)
}
\value{
TODO
}
\description{
TODO Should be managed by Bascet mapshard system, with automatic input conversion. unaligned file should be made temp and removed
}
