% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{GetBascetTempDir}
\alias{GetBascetTempDir}
\title{Get a temp directory to use; need to be created}
\usage{
GetBascetTempDir(bascet_instance)
}
\value{
TODO
}
\description{
Get a temp directory to use; need to be created
}
