% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{GetDefaultBascetInstance}
\alias{GetDefaultBascetInstance}
\title{Get default Bascet instance from global variable (bascetInstance.default)}
\usage{
GetDefaultBascetInstance()
}
\value{
A Bascet instance
}
\description{
Get default Bascet instance from global variable (bascetInstance.default)
}
