% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R
\name{GetDefaultBascetRunner}
\alias{GetDefaultBascetRunner}
\title{Get the current default runner}
\usage{
GetDefaultBascetRunner()
}
\value{
The current default runner
}
\description{
Get the current default runner
}
