% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{GetFASTQCassembledDF}
\alias{GetFASTQCassembledDF}
\title{Get a data frame for one type of FASTQ statistics across across all cells}
\usage{
GetFASTQCassembledDF(aggrFastqData, section, readnum = c(1, 2))
}
\arguments{
\item{aggrFastqData}{Aggregated FASTQC data list from BascetAggregateFASTQC}

\item{section}{Name of the FASTQC section to extract (string)}

\item{readnum}{1 or 2, for R1 and R2}
}
\value{
TODO
}
\description{
Get a data frame for one type of FASTQ statistics across across all cells
}
