% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{GetFASTQCbasicStats}
\alias{GetFASTQCbasicStats}
\title{From aggregated FASTQC data, get basic statistics for overlay on UMAP etc}
\usage{
GetFASTQCbasicStats(aggrFastqData, readnum = c(1, 2))
}
\arguments{
\item{aggrFastqData}{Aggregated FASTQ data}

\item{readnum}{1 or 2, for R1 or R2}
}
\value{
A data.frame
}
\description{
From aggregated FASTQC data, get basic statistics for overlay on UMAP etc
}
