% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{GetFASTQCpassfailStats}
\alias{GetFASTQCpassfailStats}
\title{From aggregated FASTQC data, get overall statistics for overlay on UMAP etc}
\usage{
GetFASTQCpassfailStats(aggr_fastq_data, readnum)
}
\arguments{
\item{readnum}{1 or 2, for R1 or R2}
}
\description{
TODO argument if giving many cells
}
