% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{helper_cancel_job_if_file_exists}
\alias{helper_cancel_job_if_file_exists}
\title{Create a piece of script to exit a job early if file exists}
\usage{
helper_cancel_job_if_file_exists(fvar)
}
\value{
script
}
\description{
Create a piece of script to exit a job early if file exists
}
