% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{is.bascet.file}
\alias{is.bascet.file}
\title{Check that parameter is a valid bascet file}
\usage{
is.bascet.file(f)
}
\arguments{
\item{f}{An object to test for Bascet class}
}
\description{
Check that parameter is a valid bascet file
}
