% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{is.bascet.instance}
\alias{is.bascet.instance}
\title{Check that parameter is a valid bascet instance}
\usage{
is.bascet.instance(x)
}
\arguments{
\item{x}{An object to test for BascetInstance class}
}
\description{
Check that parameter is a valid bascet instance
}
