% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{is.BascetCountMatrix}
\alias{is.BascetCountMatrix}
\title{Check that parameter is a BascetCountMatrix}
\usage{
is.BascetCountMatrix(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check that parameter is a BascetCountMatrix
}
