% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{is.existing.fasta}
\alias{is.existing.fasta}
\title{Check that parameter is a valid shard name}
\usage{
is.existing.fasta(x)
}
\arguments{
\item{x}{A file path string pointing to a FASTA file}
}
\description{
Check that parameter is a valid shard name
}
