% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{is.percent.string}
\alias{is.percent.string}
\title{Check that parameter is in the form "xxx\%"}
\usage{
is.percent.string(s)
}
\arguments{
\item{s}{A string to test for percent format (e.g. "10\%")}
}
\description{
Check that parameter is in the form "xxx\%"
}
