% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{is.positive.integer}
\alias{is.positive.integer}
\title{Check that parameter is an integer and >0}
\usage{
is.positive.integer(x)
}
\arguments{
\item{x}{A numeric value}
}
\description{
Check that parameter is an integer and >0
}
