% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{is.valid.env.variable}
\alias{is.valid.env.variable}
\title{Check that parameter is a valid environment variable name}
\usage{
is.valid.env.variable(x)
}
\arguments{
\item{x}{A string representing an environment variable name}
}
\description{
Check that parameter is a valid environment variable name
}
