% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{is.valid.memsize}
\alias{is.valid.memsize}
\title{Check that parameter is a valid memory size}
\usage{
is.valid.memsize(x)
}
\arguments{
\item{x}{A string representing a memory size (e.g. "8g")}
}
\description{
Check that parameter is a valid memory size
}
