% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{is_bam_paired_alignment}
\alias{is_bam_paired_alignment}
\title{Figure out if a BAM-file is a paired alignment or not}
\usage{
is_bam_paired_alignment(fname, bascet_instance = GetDefaultBascetInstance())
}
\arguments{
\item{fname}{Name of BAM-file}
}
\description{
Figure out if a BAM-file is a paired alignment or not
}
