% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{is_fastq}
\alias{is_fastq}
\title{Check if a file is a FASTQ file}
\usage{
is_fastq(fname)
}
\arguments{
\item{fname}{Path to file}
}
\value{
TRUE if the file is some type of FASTQ
}
\description{
Check if a file is a FASTQ file
}
