% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{isBamPairedAlignment}
\alias{isBamPairedAlignment}
\title{Figure out if a BAM-file is a paired alignment or not}
\usage{
isBamPairedAlignment(fname, bascetInstance = GetDefaultBascetInstance())
}
\arguments{
\item{fname}{Name of BAM-file}

\item{bascetInstance}{A Bascet instance}
}
\description{
Figure out if a BAM-file is a paired alignment or not
}
