% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{isPairedFastq}
\alias{isPairedFastq}
\title{Check if a file is a paired FASTQ file.
Panics if the file is not a FASTQ at all}
\usage{
isPairedFastq(fname)
}
\arguments{
\item{fname}{Path to file}
}
\value{
TRUE if the file is a paired FASTQ
}
\description{
Check if a file is a paired FASTQ file.
Panics if the file is not a FASTQ at all
}
