% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R, R/job_local.R, R/job_slurm.R
\name{JobLog}
\alias{JobLog}
\alias{JobLog,LocalJob-method}
\alias{JobLog,SlurmJob-method}
\title{Get job log}
\usage{
JobLog(job)

\S4method{JobLog}{LocalJob}(job)

\S4method{JobLog}{SlurmJob}(job)
}
\arguments{
\item{job}{A job object.}
}
\value{
Backend-specific log content, usually invisible if no log is available.
}
\description{
Get job log
}
\section{Methods (by class)}{
\itemize{
\item \code{JobLog(LocalJob)}: Local job method.

\item \code{JobLog(SlurmJob)}: SLURM job method.

}}
