% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R, R/job_local.R, R/job_slurm.R
\name{JobStatus}
\alias{JobStatus}
\alias{JobStatus,NoJob-method}
\alias{JobStatus,LocalJob-method}
\alias{JobStatus,SlurmJob-method}
\title{Get job status}
\usage{
JobStatus(job)

\S4method{JobStatus}{NoJob}(job)

\S4method{JobStatus}{LocalJob}(job)

\S4method{JobStatus}{SlurmJob}(job)
}
\arguments{
\item{job}{A job object.}
}
\value{
A data frame or backend-specific status object.
}
\description{
Get job status
}
\section{Methods (by class)}{
\itemize{
\item \code{JobStatus(NoJob)}: No-op job method.

\item \code{JobStatus(LocalJob)}: Local job method.

\item \code{JobStatus(SlurmJob)}: SLURM job method.

}}
