% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{KmcChooseKmerFeatures}
\alias{KmcChooseKmerFeatures}
\title{Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range}
\usage{
KmcChooseKmerFeatures(fname, numPick = 1000, minFreq = 0.01, maxFreq = 0.1)
}
\arguments{
\item{fname}{description}

\item{numPick}{description}

\item{minFreq}{description}

\item{maxFreq}{description}
}
\value{
List of KMERs
}
\description{
Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range
}
