% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{KmcGetHistogram}
\alias{KmcGetHistogram}
\title{Generate a histogram from a KMC3 database}
\usage{
KmcGetHistogram(fname, plot = TRUE)
}
\arguments{
\item{fname}{Full name of KMC3 database}

\item{toplot}{Decide if to plot or return raw data}
}
\value{
A ggplot object if toplot=TRUE, otherwise a data.frame
}
\description{
Generate a histogram from a KMC3 database
}
