% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{KneeplotPerSpecies}
\alias{KneeplotPerSpecies}
\title{Produce a kneeplot}
\usage{
KneeplotPerSpecies(adata, maxSpecies = NULL, showTotal = FALSE)
}
\arguments{
\item{adata}{A Seurat object with the DefaultAssay having counts per species (or similar)}

\item{maxSpecies}{Maximum number of species to show. The most abundant species will be shown first}

\item{showTotal}{Include a separate kneeplot line for the total count across all species}
}
\value{
A ggplot object
}
\description{
Produce a kneeplot
}
