% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenFindMaxTaxid}
\alias{KrakenFindMaxTaxid}
\title{For a Kraken count matrix, return consesus taxid for each cell as metadata}
\usage{
KrakenFindMaxTaxid(mat)
}
\value{
TODO
}
\description{
For a Kraken count matrix, return consesus taxid for each cell as metadata
}
