% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenSpeciesDistribution}
\alias{KrakenSpeciesDistribution}
\title{Using a KRAKEN2 count matrix, produce a "kneeplot" of species}
\usage{
KrakenSpeciesDistribution(adata, useAssay = "kraken")
}
\arguments{
\item{adata}{A Seurat object}

\item{useAssay}{Which assay to use}
}
\value{
A ggplot object
}
\description{
Using a KRAKEN2 count matrix, produce a "kneeplot" of species
}
