% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_local.R
\name{LocalInstance}
\alias{LocalInstance}
\title{Create new local runner instance}
\usage{
LocalInstance(maxcpu = "10", direct = FALSE, show_script = FALSE)
}
\value{
TODO
}
\description{
Create new local runner instance
}
