% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_local.R
\docType{class}
\name{LocalJob-class}
\alias{LocalJob-class}
\title{Local job}
\description{
Job object for a locally launched process.
}
\section{Slots}{

\describe{
\item{\code{cmd}}{Command string.}

\item{\code{proc}}{Process handle.}

\item{\code{logfile}}{Path to the job log file.}

\item{\code{arraysize}}{Number of array tasks.}
}}

