% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_local.R
\name{LocalRunner}
\alias{LocalRunner}
\title{Create new local runner instance}
\usage{
LocalRunner(maxcpu = "10", direct = TRUE, showScript = FALSE)
}
\arguments{
\item{maxcpu}{Number of CPUs to use (max)}

\item{direct}{Run jobs synchronously}

\item{showScript}{Show the script to run, for debugging purposes}
}
\value{
A runner instance
}
\description{
Create new local runner instance
}
