% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{lseq}
\alias{lseq}
\title{logarithmic spaced sequence; taken from emdbook library}
\usage{
lseq(from = 1, to = 1e+05, length.out = 6)
}
\arguments{
\item{from}{First value}

\item{to}{Last value}

\item{length.out}{Length of list}
}
\value{
List of values
}
\description{
logarithmic spaced sequence; taken from emdbook library
}
