% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{make_output_shard_names}
\alias{make_output_shard_names}
\title{Helper function: Generate suitable output filenames according to shard system
i.e. root/name.##.ext}
\usage{
make_output_shard_names(bascetRoot, outputName, ext, num_shards)
}
\description{
Helper function: Generate suitable output filenames according to shard system
i.e. root/name.##.ext
}
