% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{MapCellMultiListAsDataFrame}
\alias{MapCellMultiListAsDataFrame}
\title{Convenience function; alternative is to somehow implement as.data.frame}
\usage{
MapCellMultiListAsDataFrame(mylist)
}
\value{
TODO
}
\description{
This one puts cellID as a new column
}
\details{
FUTURE possible make this the new MapListAsDataFrame
}
