% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R
\name{new_no_job}
\alias{new_no_job}
\title{Create an empty job. It is considered to have terminated from the start}
\usage{
new_no_job()
}
\value{
A no-job
}
\description{
Create an empty job. It is considered to have terminated from the start
}
