% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{NewBascetCountMatrix}
\alias{NewBascetCountMatrix}
\title{Create new BascetCountMatrix}
\usage{
NewBascetCountMatrix(X, obs)
}
\arguments{
\item{X}{Sparse count matrix (dgCMatrix or similar)}

\item{obs}{Data frame of observation metadata, one row per cell}
}
\value{
A Bascet count matrix object
}
\description{
Create new BascetCountMatrix
}
