% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R
\docType{class}
\name{NoJob-class}
\alias{NoJob-class}
\alias{show,NoJob-method}
\title{No-op job}
\usage{
\S4method{show}{NoJob}(object)
}
\arguments{
\item{object}{A NoJob object.}
}
\description{
Placeholder job used when no asynchronous process exists.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(NoJob)}: Print a compact representation of a no-op job.

}}
\section{Slots}{

\describe{
\item{\code{evil}}{Placeholder character slot to keep the S4 class concrete.}
}}

