% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R
\docType{class}
\name{NoRunner-class}
\alias{NoRunner-class}
\title{No-op runner}
\description{
Runner backend used for dry runs and debugging.
}
\section{Slots}{

\describe{
\item{\code{showScript}}{Logical flag indicating whether scripts should be shown.}
}}

