% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R
\name{NoRunner}
\alias{NoRunner}
\title{Create new No-runner instance. This system is for testing only}
\usage{
NoRunner(show_script = FALSE)
}
\value{
A no-runner
}
\description{
Create new No-runner instance. This system is for testing only
}
