% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{parse_size_string}
\alias{parse_size_string}
\title{Parse a string with a size, such as 1g, 1m, 1k, or just 123 (bytes)}
\usage{
parse_size_string(s)
}
\arguments{
\item{s}{Size as string. If numeric, it is just returned}
}
\value{
Size in bytes, as integer in a string (long format)
}
\description{
Parse a string with a size, such as 1g, 1m, 1k, or just 123 (bytes)
}
