% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{PlotFASTQCadapterContent}
\alias{PlotFASTQCadapterContent}
\title{From aggregated FASTQC data, plot adapter content}
\usage{
PlotFASTQCadapterContent(aggr_fastq_data, readnum)
}
\arguments{
\item{readnum}{1 or 2, for R1 or R2}
}
\description{
TODO if giving many cells
}
