% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{Plot a histogram, loaded by ReadHistogram}
\usage{
PlotHistogram(dat)
}
\arguments{
\item{dat}{Histogram object from ReadHistogram}
}
\value{
A ggplot object
}
\description{
Plot a histogram, loaded by ReadHistogram
}
