% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{PlotJohnsonLindenstraussMinDim}
\alias{PlotJohnsonLindenstraussMinDim}
\title{Plot minimum number of dimensions needed to retain distance between samples}
\usage{
PlotJohnsonLindenstraussMinDim(list_eps, min_cells = 10, max_cells = 1e+06)
}
\arguments{
\item{list_eps}{List of eps to plot for}
}
\description{
https://en.wikipedia.org/wiki/Johnson–Lindenstrauss_lemma
}
\details{
(1 - eps) ||u - v||^2 < ||p(u) - p(v)||^2 < (1 + eps) ||u - v||^2
}
