% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{PrependBascetCountMatrixName}
\alias{PrependBascetCountMatrixName}
\title{Prefix names of cells in a count matrix}
\usage{
PrependBascetCountMatrixName(mat, prependName)
}
\description{
Prefix names of cells in a count matrix
}
