% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{ReadBascetCountMatrix}
\alias{ReadBascetCountMatrix}
\title{Read a count matrix as produced by Bascet (hdf5 format).
This can be output from both BascetQueryFq and BascetCountChrom}
\usage{
ReadBascetCountMatrix(bascetRoot, inputName, verbose = FALSE)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
BascetCountMatrix
}
\description{
Read a count matrix as produced by Bascet (hdf5 format).
This can be output from both BascetQueryFq and BascetCountChrom
}
