% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{ReadBascetKrakenMatrix}
\alias{ReadBascetKrakenMatrix}
\title{Read a KRAKEN2 count matrix as produced by Bascet (hdf5 format)}
\usage{
ReadBascetKrakenMatrix(bascetRoot, inputName = "kraken")
}
\arguments{
\item{fname}{Full name of the HDF5 count matrix file}
}
\value{
Counts as a sparseMatrix
}
\description{
Read a KRAKEN2 count matrix as produced by Bascet (hdf5 format)
}
