% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{ReadBascetKrakenMatrix_one}
\alias{ReadBascetKrakenMatrix_one}
\title{Internal function, to load a single kraken matrix}
\usage{
ReadBascetKrakenMatrix_one(fname)
}
\description{
Internal function, to load a single kraken matrix
}
