% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{ReadCellSNPmatrix}
\alias{ReadCellSNPmatrix}
\title{Read a count matrix as produced by CellSNP, but as shards}
\usage{
ReadCellSNPmatrix(bascetRoot, inputName, listCells = NULL, verbose = FALSE)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{listCells}{Optional list of cells to extract}

\item{verbose}{Show process status for debugging purposes}
}
\value{
Count matrix as sparseMatrix
}
\description{
Read a count matrix as produced by CellSNP, but as shards
}
