% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{ReadHistogram}
\alias{ReadHistogram}
\title{Read the count histogram associated with a Bascet.
Not all Bascets have one, but it is typically produced after debarcoding}
\usage{
ReadHistogram(
  bascetRoot,
  inputName,
  bascetInstance = GetDefaultBascetInstance(),
  verbose = TRUE
)
}
\arguments{
\item{bascetInstance}{A Bascet instance}

\item{minCount}{Minimum count for a cell to be included. Setting this can have a great impact on speed}
}
\value{
Histogram as a data.frame
}
\description{
Read the count histogram associated with a Bascet.
Not all Bascets have one, but it is typically produced after debarcoding
}
