% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{rownames,BascetCountMatrix-method}
\alias{rownames,BascetCountMatrix-method}
\title{Row names of a BascetCountMatrix (cell names)}
\usage{
\S4method{rownames}{BascetCountMatrix}(x)
}
\arguments{
\item{x}{A BascetCountMatrix object}
}
\value{
Character vector of row names
}
\description{
Row names of a BascetCountMatrix (cell names)
}
