% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{rownames<-,BascetCountMatrix-method}
\alias{rownames<-,BascetCountMatrix-method}
\title{Set row names of a BascetCountMatrix (cell names)}
\usage{
\S4method{rownames}{BascetCountMatrix}(x) <- value
}
\arguments{
\item{x}{A BascetCountMatrix object}

\item{value}{Character vector of new row names}
}
\value{
The modified BascetCountMatrix
}
\description{
Set row names of a BascetCountMatrix (cell names)
}
