% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{rowSums,BascetCountMatrix-method}
\alias{rowSums,BascetCountMatrix-method}
\title{Row sums of a BascetCountMatrix (counts per cell)}
\usage{
\S4method{rowSums}{BascetCountMatrix}(x)
}
\arguments{
\item{x}{A BascetCountMatrix object}
}
\value{
Named numeric vector of row sums
}
\description{
Row sums of a BascetCountMatrix (counts per cell)
}
