% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R, R/job_local.R, R/job_slurm.R
\name{RunJob}
\alias{RunJob}
\alias{RunJob,NoRunner-method}
\alias{RunJob,LocalRunner-method}
\alias{RunJob,SlurmRunner-method}
\title{Run a job}
\usage{
RunJob(runner, jobname, bascetInstance, cmd, arraysize)

\S4method{RunJob}{NoRunner}(runner, jobname, bascetInstance, cmd, arraysize)

\S4method{RunJob}{LocalRunner}(runner, jobname, bascetInstance, cmd, arraysize)

\S4method{RunJob}{SlurmRunner}(runner, jobname, bascetInstance, cmd, arraysize)
}
\arguments{
\item{runner}{Runner object used to execute the job.}

\item{jobname}{Name for the submitted job.}

\item{bascetInstance}{Bascet instance that provides runtime settings.}

\item{cmd}{JobScript object to execute.}

\item{arraysize}{Number of array tasks.}
}
\value{
A job object, or a no-op job for synchronous/no-runner backends.
}
\description{
Submit a job script with a runner backend.
}
\section{Methods (by class)}{
\itemize{
\item \code{RunJob(NoRunner)}: No-op runner method.

\item \code{RunJob(LocalRunner)}: Local runner method.

\item \code{RunJob(SlurmRunner)}: SLURM runner method.

}}
