% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{safeDownloadMD5}
\alias{safeDownloadMD5}
\title{Download a file, check MD5 to ensure success}
\usage{
safeDownloadMD5(url, file)
}
\value{
nothing; panics if the download fails
}
\description{
Download a file, check MD5 to ensure success
}
