% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscript_make_bash_array}
\alias{shellscript_make_bash_array}
\title{Helper function: Create array of values in bash scripts}
\usage{
shellscript_make_bash_array(variable, vals)
}
\description{
Helper function: Create array of values in bash scripts
}
