% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscriptMakeBashArray}
\alias{shellscriptMakeBashArray}
\title{Helper function: Create array of values in bash scripts}
\usage{
shellscriptMakeBashArray(variable, vals)
}
\description{
Helper function: Create array of values in bash scripts
}
