% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscriptMakeCommalist}
\alias{shellscriptMakeCommalist}
\title{Helper function, taking a list of elements such as \link{a,b}, and returning "a,b"}
\usage{
shellscriptMakeCommalist(f)
}
\arguments{
\item{f}{List of elements}
}
\value{
Formatted list
}
\description{
Helper function, taking a list of elements such as \link{a,b}, and returning "a,b"
}
