% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscriptMakeOneFileExpander}
\alias{shellscriptMakeOneFileExpander}
\title{Helper function that takes content of a list and generates a BASH script
that stores the content in a temporary file during execution}
\usage{
shellscriptMakeOneFileExpander(tmpname, list_lines)
}
\arguments{
\item{tmpname}{Name of environment variable in which the name of the file will be stored}

\item{list_lines}{Content to write to the file}
}
\value{
BASH script content for the expander
}
\description{
Helper function that takes content of a list and generates a BASH script
that stores the content in a temporary file during execution
}
