% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscriptSplitArrayIntoListRandomly}
\alias{shellscriptSplitArrayIntoListRandomly}
\title{Helper function to take an array of elements and split it randomly into a number
of subset lists}
\usage{
shellscriptSplitArrayIntoListRandomly(arr, num_divide)
}
\arguments{
\item{arr}{Array of elements to subdivide}

\item{num_divide}{Number of subdivions}
}
\value{
List of arrays
}
\description{
Helper function to take an array of elements and split it randomly into a number
of subset lists
}
