% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_slurm.R
\name{SlurmInstance}
\alias{SlurmInstance}
\title{Create a runner that submits jobs to SLURM}
\usage{
SlurmInstance(
  settings = NULL,
  ncpu = NULL,
  partition = NULL,
  account = NULL,
  time = NULL,
  prepend = NULL,
  mem = NULL
)
}
\arguments{
\item{settings}{Default settings to override; can be NULL}

\item{ncpu}{description}

\item{partition}{description}

\item{account}{description}

\item{time}{The time the job is allowed to run, e.g. "0-72:00:00"}

\item{prepend}{description}

\item{mem}{description}
}
\value{
A SLURM runner
}
\description{
Create a runner that submits jobs to SLURM
}
