% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_slurm.R
\docType{class}
\name{SlurmJob-class}
\alias{SlurmJob-class}
\title{SLURM job}
\description{
Job object for a submitted SLURM array job.
}
\section{Slots}{

\describe{
\item{\code{pid}}{SLURM job identifier.}

\item{\code{cmd}}{Submitted command.}

\item{\code{jobname}}{Job name.}

\item{\code{arraysize}}{Number of array tasks.}
}}

