% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{SpeciesCorrMatrix}
\alias{SpeciesCorrMatrix}
\title{TODO is this function used? export}
\usage{
SpeciesCorrMatrix(adata)
}
\arguments{
\item{adata}{A Seurat object with the DefaultAssay having counts per species}
}
\value{
A ggplot object
}
\description{
TODO is this function used? export
}
