% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{template_BascetFunction}
\alias{template_BascetFunction}
\title{Function template, where basic parameter documentation can be obtained from}
\usage{
template_BascetFunction(
  bascetRoot,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{Configuration for how to run the Bascet Rust API}

\item{num_output_shards}{Number of output shards, i.e., how many output files to split the data into? (>=1)}

\item{includeCells}{List of cells to process}

\item{bascetFile}{Handle for an opened Bascet file}

\item{cellID}{ID of the cell (string)}

\item{verbose}{description}
}
\value{
A job to be executed
}
\description{
Function template, where basic parameter documentation can be obtained from
}
