% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{TestBascetInstance}
\alias{TestBascetInstance}
\title{Check if a Bascet instance works}
\usage{
TestBascetInstance(bascet_instance)
}
\arguments{
\item{bascet_instance}{Bascet instance}
}
\value{
"ok" if the instance works; panic otherwise
}
\description{
Check if a Bascet instance works
}
