% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R, R/job_slurm.R
\name{WaitForJob}
\alias{WaitForJob}
\alias{WaitForJob,NoJob-method}
\alias{WaitForJob,SlurmJob-method}
\title{Wait for a job to finish}
\usage{
WaitForJob(job)

\S4method{WaitForJob}{NoJob}(job)

\S4method{WaitForJob}{SlurmJob}(job)
}
\arguments{
\item{job}{A job object.}
}
\value{
Invisibly returns when the job has completed.
}
\description{
Wait for a job to finish
}
\section{Methods (by class)}{
\itemize{
\item \code{WaitForJob(NoJob)}: No-op job method.

\item \code{WaitForJob(SlurmJob)}: SLURM job method.

}}
