% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{aggr.amrfinder}
\alias{aggr.amrfinder}
\title{Callback function for aggregating ABRicate data for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.amrfinder(bascetFile, cellID, bascetInstance)
}
\arguments{
\item{bascetFile}{An opened Bascet file}

\item{cellID}{Cell ID}

\item{bascetInstance}{A Bascet instance}
}
\value{
Data for one cell
}
\description{
Callback function for aggregating ABRicate data for each cell.
To be called from BascetAggregateMap
}
