% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_functions.R
\name{aggr.ariba}
\alias{aggr.ariba}
\title{Callback function for aggregating ARIBA data for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.ariba(bascetFile, cellID, bascet_instance)
}
\value{
TODO
}
\description{
Callback function for aggregating ARIBA data for each cell.
To be called from BascetAggregateMap
}
