% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{aggr.fastqc}
\alias{aggr.fastqc}
\title{Callback function for aggregating FASTQC data for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.fastqc(bascetFile, cellID, bascet_instance)
}
\value{
TODO
}
\description{
Callback function for aggregating FASTQC data for each cell.
To be called from BascetAggregateMap
}
