% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_functions.R
\name{aggr.minhash}
\alias{aggr.minhash}
\title{Callback function for aggregating min-hashes for each cell.
To be called from BascetAggregateMap}
\usage{
aggr.minhash(bascetFile, cellID, bascetInstance)
}
\arguments{
\item{bascetFile}{Bascet file handle}

\item{cellID}{ID of cell to process}

\item{bascetInstance}{A Bascet instance}
}
\value{
Minhash data (minhash.txt) for each cell
}
\description{
Callback function for aggregating min-hashes for each cell.
To be called from BascetAggregateMap
}
