% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{aggr.quast}
\alias{aggr.quast}
\title{Callback function for aggregating QUAST data.
To be called from BascetAggregateMap}
\usage{
aggr.quast(bascetFile, cellID, bascetInstance)
}
\value{
QUAST data for each cell
}
\description{
Callback function for aggregating QUAST data.
To be called from BascetAggregateMap
}
