% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{AggregateMinhashes}
\alias{AggregateMinhashes}
\title{Aggregate frequency of minhashes across cells}
\usage{
AggregateMinhashes(bascetRoot, inputName = "minhash")
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{description}
}
\value{
TODO
}
\description{
Aggregate frequency of minhashes across cells
}
