% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_functions.R
\name{AggregateMinhashes}
\alias{AggregateMinhashes}
\title{Aggregate frequency of minhashes across cells}
\usage{
AggregateMinhashes(bascetRoot, inputName = "minhash", bascetInstance)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{description}

\item{bascetInstance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
Aggregate frequency of minhashes across cells
}
