% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateAbricate}
\alias{BascetAggregateAbricate}
\title{Aggregate data from Abricate
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateAbricate(
  bascetRoot,
  inputName = "abricate",
  include_cells = NULL,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\description{
Aggregate data from Abricate
This is a thin wrapper around BascetAggregateMap
}
