% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateAMRfinder}
\alias{BascetAggregateAMRfinder}
\title{Aggregate data from AMRfinder
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateAMRfinder(
  bascetRoot,
  inputName = "AMRfinder",
  include_cells = NULL,
  get_column = "Element.symbol",
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\description{
Aggregate data from AMRfinder
This is a thin wrapper around BascetAggregateMap
}
