% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateAriba}
\alias{BascetAggregateAriba}
\title{Aggregate data from Ariba
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateAriba(
  bascetRoot,
  inputName = "ariba",
  include_cells = NULL,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\description{
Aggregate data from Ariba
This is a thin wrapper around BascetAggregateMap
}
