% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateFASTQC}
\alias{BascetAggregateFASTQC}
\title{Aggregate data from FASTQC
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateFASTQC(
  bascetRoot,
  inputName = "fastqc",
  include_cells = NULL,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\description{
Aggregate data from FASTQC
This is a thin wrapper around BascetAggregateMap
}
