% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateGECCO}
\alias{BascetAggregateGECCO}
\title{Aggregate data from GECCO
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateGECCO(
  bascetRoot,
  inputName = "gecco",
  includeCells = NULL,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\description{
Aggregate data from GECCO
This is a thin wrapper around BascetAggregateMap
}
