% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{BascetAggregateMap}
\alias{BascetAggregateMap}
\title{Aggregate data from previous Map call}
\usage{
BascetAggregateMap(
  bascetRoot,
  bascetName,
  aggrFunction,
  includeCells = NULL,
  showProgress = TRUE,
  verbose = FALSE,
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{bascetName}{Name of input shard}

\item{aggrFunction}{Function to use for extracting simplified data for cell}

\item{includeCells}{Cells to aggregate}

\item{showProgress}{Show progress bar}

\item{verbose}{Show debug output}
}
\value{
Aggregated data
}
\description{
Aggregate data from previous Map call
}
