% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetAggregateQUAST}
\alias{BascetAggregateQUAST}
\title{Aggregate data from QUAST
This is a thin wrapper around BascetAggregateMap}
\usage{
BascetAggregateQUAST(
  bascetRoot,
  inputName = "quast",
  includeCells = NULL,
  verbose = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{verbose}{Show debug output}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
Aggregated data
}
\description{
Aggregate data from QUAST
This is a thin wrapper around BascetAggregateMap
}
