% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetAlignmentToBigwig}
\alias{BascetAlignmentToBigwig}
\title{Generate a bigwig out of all reads in a sorted BAM. Note that the caller
is responsible for sorting the BAM first}
\usage{
BascetAlignmentToBigwig(
  bascetRoot,
  inputName = "aligned",
  outputName = "pileup",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard (BAM-file)}

\item{outputName}{Name of output shard (BIGWIG-file)}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A runner job (details depends on runner)
}
\description{
This function is mainly for QC purposes. It uses bamCoverage from deepTools
apt install python3-deeptools
}
