% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetAlignmentToBigwig}
\alias{BascetAlignmentToBigwig}
\title{Generate a bigwig out of all reads in a sorted BAM. Note that the caller
is responsible for sorting the BAM first}
\usage{
BascetAlignmentToBigwig(
  bascetRoot,
  inputName = "aligned",
  outputName = "pileup",
  overwrite = FALSE,
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{outputName}{Name of output file (BIGWIG-file)}
}
\description{
This function is mainly for QC purposes. It uses bamCoverage from deepTools
apt install python3-deeptools
}
