% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetAlignToReference}
\alias{BascetAlignToReference}
\title{Align from FASTQ, generate sorted and indexed BAM file}
\usage{
BascetAlignToReference(
  bascetRoot,
  useReference,
  numLocalThreads = 1,
  inputName = "asfq",
  outputNameBAMunsorted = "unsorted_aligned",
  outputNameBAMsorted = "aligned",
  doSort = TRUE,
  overwrite = FALSE,
  aligner = "BWA",
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{useReference}{Name of the BWA reference to use}

\item{numLocalThreads}{Number of threads to use for each runner}

\item{inputName}{Name of input shard}

\item{outputNameBAMunsorted}{Name of unsorted BAMs}

\item{outputNameBAMsorted}{Name of sorted BAMs (if generated)}

\item{doSort}{Whether to sort the output or not}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\description{
Align from FASTQ, generate sorted and indexed BAM file
}
