% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetAlignToReference}
\alias{BascetAlignToReference}
\title{Align from FASTQ, generate sorted and indexed BAM file}
\usage{
BascetAlignToReference(
  bascetRoot,
  useReference,
  numLocalThreads = 1,
  inputName = "asfq",
  outputNameBAMunsorted = "unsorted_aligned",
  outputNameBAMsorted = "aligned",
  do_sort = TRUE,
  overwrite = FALSE,
  aligner = "BWA",
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\arguments{
\item{useReference}{Name of the BWA reference to use}

\item{numLocalThreads}{Number of threads to use for each runner}

\item{outputNameBAMunsorted}{Name of unsorted BAMs}

\item{outputNameBAMsorted}{Name of sorted BAMs (if generated)}

\item{do_sort}{Whether to sort the output or not}
}
\description{
TODO Should be managed by Bascet mapshard system, with automatic input conversion. unaligned file should be made temp and removed
}
