% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetBam2Fragments}
\alias{BascetBam2Fragments}
\title{Take aligned BAM file and produce Fragments.tsv.gz, compatible with Signac ATAC-seq style analysis}
\usage{
BascetBam2Fragments(
  bascetRoot,
  inputName = "aligned",
  outputName = "fragments",
  overwrite = FALSE,
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
A job, producing a type of Fragments.tsv.gz
}
\description{
Take aligned BAM file and produce Fragments.tsv.gz, compatible with Signac ATAC-seq style analysis
}
