% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{BascetComputeCountSketch}
\alias{BascetComputeCountSketch}
\title{Compute count sketch for each cell.
This is a thin wrapper around BascetMapCell}
\usage{
BascetComputeCountSketch(
  bascetRoot,
  inputName = "filtered",
  outputName = "countsketch",
  overwrite = FALSE,
  maxReads = 1e+05,
  kmerSize = 31,
  sketchSize = 5000,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{maxReads}{Max number of reads to sample per cell}

\item{kmerSize}{Size of the KMER to hash}

\item{sketchSize}{Number of dimensions to reduce to}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Compute count sketch for each cell.
This is a thin wrapper around BascetMapCell
}
