% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetComputeMinhash}
\alias{BascetComputeMinhash}
\title{Compute minhashes for each cell.
This is a thin wrapper around BascetMapCell}
\usage{
BascetComputeMinhash(
  bascetRoot,
  inputName = "filtered",
  outputName = "minhash",
  overwrite = FALSE,
  maxReads = 1e+05,
  kmerSize = 31,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{maxReads}{The maximum number of reads per cell to sample}

\item{kmerSize}{The KMER size for the hashing}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Compute minhashes for each cell.
This is a thin wrapper around BascetMapCell
}
