% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetCountChrom}
\alias{BascetCountChrom}
\title{From aligned BAM file, compute counts per chromosome}
\usage{
BascetCountChrom(
  bascetRoot,
  inputName = "aligned",
  outputName = "chromcount",
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
From aligned BAM file, compute counts per chromosome
}
