% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetCountFeature}
\alias{BascetCountFeature}
\title{From aligned BAM file, compute counts per feature}
\usage{
BascetCountFeature(
  bascetRoot,
  inputName = "aligned",
  outputName = "featurecount",
  gffFile,
  useFeature = "gene",
  attrGeneId = "gene_id",
  attrGeneName = "name",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\value{
A job, executing the counting
}
\description{
From aligned BAM file, compute counts per feature
}
