% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetDumpContigs}
\alias{BascetDumpContigs}
\title{Store all contigs in an output directory, as cell_id.fa}
\usage{
BascetDumpContigs(
  bascetRoot,
  inputName = "skesa",
  listCells,
  outputDir,
  bascetInstance
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard (should hold contigs)}

\item{listCells}{List of cells to extract}

\item{outputDir}{Directory to store FASTA in}

\item{bascetInstance}{A Bascet instance}
}
\value{
Nothing
}
\description{
Store all contigs in an output directory, as cell_id.fa
}
