% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetFilterAlignment}
\alias{BascetFilterAlignment}
\title{Filter an alignment (BAM-file).}
\usage{
BascetFilterAlignment(
  bascetRoot,
  numLocalThreads = NULL,
  inputName,
  outputName,
  keepMapped = FALSE,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{numLocalThreads}{Number of threads to use. Default is the maximum, taken from runner settings}

\item{inputName}{Name of input shards (BAM-file format)}

\item{outputName}{Name of output shards (BAM-file format)}

\item{keepMapped}{Keep the mapped reads (TRUE) or unmapped (FALSE)}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
This is typically used to either remove host DNA, or keep reads mapping to a known reference.
If the BAM-file has paired reads then BOTH reads need to be mapped (flag 0x2); otherwise (flag 0x4)
}
