% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{BascetGatherCountSketch}
\alias{BascetGatherCountSketch}
\title{Gather all count sketches into a single count sketch matrix}
\usage{
BascetGatherCountSketch(
  bascetRoot,
  inputName = "countsketch",
  outputName = "countsketch_mat.csv",
  includeCells = NULL,
  overwrite = FALSE,
  numLocalThreads = NULL,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output file}

\item{includeCells}{List of cells to process}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{numLocalThreads}{Number of threads to use per job. Default is the number from the runner}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance

TODO produce a binary file format instead; gather files upon loading?}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Gather all count sketches into a single count sketch matrix
}
