% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoding.R
\name{BascetGetRaw}
\alias{BascetGetRaw}
\title{Extract barcodes and trim input raw FASTQ}
\usage{
BascetGetRaw(
  bascetRoot,
  rawmeta,
  maxShardSize = "200g",
  outputName = "debarcoded",
  chemistry = c("atrandi-wgs", "atrandi-wgslr", "atrandi-rnaseq", "parse-bio"),
  subchemistry = NULL,
  barcodeTolerance = NULL,
  numThreads = NULL,
  numReadThreads = NULL,
  numDebarcodeThreads = NULL,
  numSortingThreads = NULL,
  numMergeSortingThreads = NULL,
  numWriteThreads = NULL,
  numCompressThreads = NULL,
  totalMem = NULL,
  streamBufferSize = NULL,
  sortBufferSize = NULL,
  compressBufferSize = NULL,
  compressRawBufferSize = NULL,
  compressionLevel = NULL,
  numMergeStreams = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{rawmeta}{Metadata for the raw FASTQ input files. See DetectRawFileMeta}

\item{maxShardSize}{Estimated maximum size of output shard. Can be set higher but as sorting is also performed during sharding, it can be overall more efficient to only do partial sorting during this command}

\item{outputName}{Name output files: Debarcoded reads}

\item{chemistry}{The type of data to be parsed}

\item{barcodeTolerance}{Optional: Number of mismatches allowed in the barcode for it to still be considered valid}

\item{numThreads}{Number of threads to use per job. Default is the number from the runner}

\item{numReadThreads}{Advanced setting: Number of threads for reading}

\item{numDebarcodeThreads}{Advanced setting: Number of threads for debarcoding}

\item{numSortingThreads}{Advanced setting: Number of threads for sorting, first phase}

\item{numMergeSortingThreads}{Advanced setting: Number of threads for sorting, second phase}

\item{numWriteThreads}{Advanced setting: Number of threads for writing}

\item{numCompressThreads}{Advanced setting: Number of threads for compressing}

\item{totalMem}{Total memory to allocate}

\item{streamBufferSize}{Advanced setting: Stream buffer size (fraction, given as e.g. "10\%")}

\item{sortBufferSize}{Advanced setting: Sort buffer size (fraction, given as e.g. "10\%")}

\item{compressionLevel}{Advanced setting: Compression level (0..12)}

\item{overwrite}{}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}

\item{pageBufferSize}{Advanced setting: Page buffer size (fraction, given as e.g. "10\%")}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Extract barcodes and trim input raw FASTQ
}
