% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoding.R
\name{BascetGetRaw}
\alias{BascetGetRaw}
\title{Extract barcodes and trim input raw FASTQ}
\usage{
BascetGetRaw(
  bascetRoot,
  rawmeta,
  maxShardSize = "50g",
  outputName = "debarcoded",
  outputNameIncomplete = "incomplete_reads",
  chemistry = c("atrandi-wgs", "atrandi-rnaseq", "parse-bio"),
  subchemistry = NULL,
  barcodeTolerance = NULL,
  numLocalThreads = NULL,
  bufferSize = 4000,
  sortBufferSize = 4000,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{rawmeta}{Metadata for the raw FASTQ input files. See DetectRawFileMeta}

\item{maxShardSize}{Estimated maximum size of output shard. Can be set higher but as sorting is also performed during sharding, it can be overall more efficient to only do partial sorting during this command}

\item{outputName}{Name output files: Debarcoded reads}

\item{outputNameIncomplete}{Name of output files: Reads that could not be parsed}

\item{chemistry}{The type of data to be parsed}

\item{barcodeTolerance}{Optional: Number of mismatches allowed in the barcode for it to still be considered valid}

\item{numLocalThreads}{Number of threads to use per job. Default is the number from the runner}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Extract barcodes and trim input raw FASTQ
}
