% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetGetRaw}
\alias{BascetGetRaw}
\title{Generate BAM with barcodes from input raw FASTQ}
\usage{
BascetGetRaw(
  bascetRoot,
  rawmeta,
  outputName = "debarcoded",
  outputNameIncomplete = "incomplete_reads",
  chemistry = "atrandi_wgs",
  subchemistry = NULL,
  barcodeTolerance = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{rawmeta}{Metadata for the raw FASTQ input files. See DetectRawFileMeta}

\item{outputName}{Name output files: Debarcoded reads}

\item{outputNameIncomplete}{Name of output files: Reads that could not be parsed}

\item{chemistry}{The type of data to be parsed}

\item{barcodeTolerance}{Optional: Number of mismatches allowed in the barcode for it to still be considered valid}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\description{
Generate BAM with barcodes from input raw FASTQ
}
