% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeBWA}
\alias{BascetIndexGenomeBWA}
\title{Index a genome using BWA such that it can be used for alignment}
\usage{
BascetIndexGenomeBWA(
  genomeFile,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{genomeFile}{Name of FASTA file holding genome sequence}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\description{
TODO: could check if genome is indexed already
}
