% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeBWA}
\alias{BascetIndexGenomeBWA}
\title{Index a genome using BWA such that it can be used for alignment}
\usage{
BascetIndexGenomeBWA(
  bascetRoot,
  genomeFile,
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{genomeFile}{Name of FASTA file}
}
\description{
Index a genome using BWA such that it can be used for alignment
}
