% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeSTAR}
\alias{BascetIndexGenomeSTAR}
\title{Index a genome using STAR such that it can be used for alignment}
\usage{
BascetIndexGenomeSTAR(
  fastaFile,
  gtfFile,
  outDir,
  numThreads = NULL,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{fastaFile}{Name of FASTA file holding genome sequence}

\item{gtfFile}{GFF file holding genome annotation}

\item{outDir}{A directory in which to store the index. This directory will be created}

\item{numThreads}{The number of threads to use for the STAR index, for each runner. Default is the maximum, taken from runner settings}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Index a genome using STAR such that it can be used for alignment
}
