% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetIndexGenomeSTAR}
\alias{BascetIndexGenomeSTAR}
\title{Index a genome using STAR such that it can be used for alignment}
\usage{
BascetIndexGenomeSTAR(
  bascetRoot,
  fastaFile,
  gtfFile,
  outDir,
  numLocalThreads = 10,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{gtfFile}{description}

\item{outDir}{A directory in which to store the output genome. This directory will be created}

\item{genomeFile}{Name of FASTA file}
}
\description{
Index a genome using STAR such that it can be used for alignment
}
