% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetListFilesForCell}
\alias{BascetListFilesForCell}
\title{List files for a cell in a Bascet}
\usage{
BascetListFilesForCell(
  bascetFile,
  cellID,
  bascet_instance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetFile}{Bascet file object}

\item{cellID}{Name of the cell}

\item{bascet_instance}{A Bascet instance}
}
\value{
A data.frame with list of all the files
}
\description{
This can be made faster by, e.g., once and for all reading the location of
all objects in the file
}
