% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{BascetLoadCountSketchMatrix}
\alias{BascetLoadCountSketchMatrix}
\title{Load count sketch matrix as Seurat object}
\usage{
BascetLoadCountSketchMatrix(bascetRoot, inputName = "countsketch_mat.csv")
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of countsketch matrix file}
}
\value{
A Seurat object holding the sketch as a reduction
}
\description{
Load count sketch matrix as Seurat object
}
