% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{BascetLoadCountSketchMatrix}
\alias{BascetLoadCountSketchMatrix}
\title{Load count sketch matrix as Seurat object}
\usage{
BascetLoadCountSketchMatrix(bascetRoot, inputName = "countsketch_mat.csv")
}
\value{
A seurat object
}
\description{
Load count sketch matrix as Seurat object
}
