% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{BascetMapCell}
\alias{BascetMapCell}
\title{Call a MAP function for all cells}
\usage{
BascetMapCell(
  bascetRoot,
  withfunction,
  inputName,
  outputName,
  args = list(),
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{args}{List of arguments (key,value) to provide to the script}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{TODO override args docs
}\if{html}{\out{</div>}}
}
