% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R, R/zorn_aggr.R
\name{BascetMapCell}
\alias{BascetMapCell}
\title{Call a function for all cells}
\usage{
BascetMapCell(
  bascetRoot,
  withfunction,
  inputName,
  outputName,
  runner,
  bascet_instance = bascet_instance.default
)

BascetMapCell(
  bascetRoot,
  withfunction,
  inputName,
  outputName,
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
Call a function for all cells

Call a function for all cells
}
