% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{BascetMapCell}
\alias{BascetMapCell}
\title{Call a MAP function for all cells}
\usage{
BascetMapCell(
  bascetRoot,
  withfunction,
  inputName,
  outputName,
  args = list(),
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{withfunction}{Function to apply}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{args}{List of arguments (key,value) to provide to the script}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Call a MAP function for all cells
}
