% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetMapCellAbricate}
\alias{BascetMapCellAbricate}
\title{Run Abricate on contigs of all cells.
This is a thin wrapper around BascetMapCell}
\usage{
BascetMapCellAbricate(
  bascetRoot,
  inputName = "contigs",
  outputName = "abricate",
  db = "ncbi",
  args = list(),
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{db}{Name of ABRicate database to use (string)}

\item{args}{Additional arguments passed to Bascet (list)}

\item{overwrite}{Whether to overwrite existing output (logical)}

\item{runner}{A runner object (e.g. LocalRunner, SlurmRunner)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Run Abricate on contigs of all cells.
This is a thin wrapper around BascetMapCell
}
\seealso{
\code{\link{BascetMapCell}}
}
