% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetMapCellAMRfinder}
\alias{BascetMapCellAMRfinder}
\title{Run AMRfinder on contigs of all cells.
This is a thin wrapper around BascetMapCell}
\usage{
BascetMapCellAMRfinder(
  bascetRoot,
  inputName = "contigs",
  outputName = "AMRfinder",
  db,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{db}{description}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
Run AMRfinder on contigs of all cells.
This is a thin wrapper around BascetMapCell
}
