% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{BascetMapCellGECCO}
\alias{BascetMapCellGECCO}
\title{Run GEECCO on contigs of all cells.
This is a thin wrapper around BascetMapCell}
\usage{
BascetMapCellGECCO(
  bascetRoot,
  inputName = "contigs",
  outputName = "gecco",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}

\item{db}{description}
}
\value{
TODO
}
\description{
Run GEECCO on contigs of all cells.
This is a thin wrapper around BascetMapCell
}
