% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetMapTransform}
\alias{BascetMapTransform}
\title{Transform data}
\usage{
BascetMapTransform(
  bascetRoot,
  inputName,
  outputName,
  num_divide = 1,
  num_merge = 1,
  out_format = "tirp.gz",
  includeCells = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascet_instance = GetDefaultBascetInstance()
)
}
\arguments{
\item{num_divide}{description}

\item{num_merge}{description}

\item{out_format}{description}
}
\description{
This command enables
\itemize{
\item subsetting to a list of cells
\item converting between file formats
\item merging shards
\item dividing shards
}
}
