% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetMapTransform}
\alias{BascetMapTransform}
\title{Transform data}
\usage{
BascetMapTransform(
  bascetRoot,
  inputName,
  outputName,
  numDivide = 1,
  numMerge = 1,
  outFormat = "tirp.gz",
  includeCells = NULL,
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{numDivide}{Must be >=1; if more than 1, divide each input shard into number of outputs given here}

\item{numMerge}{Must be >=1; if more than 1, merge this number of input shards into one}

\item{outFormat}{Extension for the output files}

\item{includeCells}{List of cells to include, or NULL if to include all}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
This command enables
\itemize{
\item subsetting to a list of cells
\item converting between file formats
\item merging shards
\item dividing shards
}
}
