% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{BascetQueryFq}
\alias{BascetQueryFq}
\title{Build count table from FASTQ reads and a list of selected kmers}
\usage{
BascetQueryFq(
  bascetRoot,
  inputName = "filtered",
  outputName = "kmer_counts",
  useKMERs,
  max_reads = 1e+06,
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{useKMERs}{description}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}
}
\value{
TODO
}
\description{
Build count table from FASTQ reads and a list of selected kmers
}
