% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetReadFile}
\alias{BascetReadFile}
\title{Read one file from a Bascet}
\usage{
BascetReadFile(
  bascetFile,
  cellID,
  filename,
  as = c("tempfile"),
  bascet_instance = bascet_instance.default,
  verbose = FALSE
)
}
\value{
TODO
}
\description{
This can be made faster by, e.g., once and for all reading the location of
all objects in the file
}
