% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{BascetReadFile}
\alias{BascetReadFile}
\title{Read one file from a Bascet}
\usage{
BascetReadFile(
  bascetFile,
  cellID,
  filename,
  as = c("tempfile"),
  bascet_instance = bascet_instance.default,
  verbose = FALSE
)
}
\arguments{
\item{bascetFile}{Handle for an opened Bascet file}

\item{cellID}{ID of the cell (string)}

\item{filename}{description}

\item{as}{description}

\item{bascet_instance}{Configuration for how to run the Bascet Rust API}

\item{verbose}{description}
}
\value{
Name of a temporary file where the read content is stored
}
\description{
TODO This can be made faster by, e.g., once and for all reading the location of
all objects in the file
}
