% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{BascetRunCellSNP}
\alias{BascetRunCellSNP}
\title{Align from FASTQ, generate sorted and indexed BAM file}
\usage{
BascetRunCellSNP(
  bascetRoot,
  numLocalThreads = 1,
  inputName = "aligned",
  outputName = "cellsnp",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{numLocalThreads}{Number of threads to use for each runner}

\item{inputName}{xx}

\item{outputName}{xx}

\item{useReference}{Name of the BWA reference to use}
}
\description{
TODO Should be managed by Bascet mapshard system, with automatic input conversion. unaligned file should be made temp and removed
}
