% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{BascetRunFASTP}
\alias{BascetRunFASTP}
\title{Run FASTP for each cell. Input must be in FASTQ file format}
\usage{
BascetRunFASTP(
  bascetRoot,
  numLocalThreads,
  inputName = "asfq",
  outputName = "fastp",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{numLocalThreads}{Number of threads to use for FASTP. TODO: autodetect? take from runner? (fastp has no good defaults)}

\item{inputName}{Name of input shard}

\item{outputName}{Name of output shard}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Run FASTP for each cell. Input must be in FASTQ file format
}
