% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{BascetRunKraken}
\alias{BascetRunKraken}
\title{Run KRAKEN2 for each cell}
\usage{
BascetRunKraken(
  bascetRoot,
  useKrakenDB = "/data/henlab/kraken/standard-8",
  numLocalThreads = 1,
  inputName = "asfq",
  outputName = "kraken_out",
  runner,
  bascet_instance = bascet_instance.default
)
}
\arguments{
\item{useKrakenDB}{Path to KRAKEN2 database}
}
\description{
Run KRAKEN2 for each cell
}
