% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{BascetRunKraken}
\alias{BascetRunKraken}
\title{Run KRAKEN2 for each cell.
Then produce a count matrix of taxonomy IDs from the output}
\usage{
BascetRunKraken(
  bascetRoot,
  useKrakenDB = NULL,
  numThreads = NULL,
  inputName = "filtered",
  outputRawName = "kraken_raw",
  outputMatrixName = "kraken_mat",
  overwrite = FALSE,
  runner = GetDefaultBascetRunner(),
  bascetInstance = GetDefaultBascetInstance()
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{useKrakenDB}{Path to KRAKEN2 database}

\item{numThreads}{Number of threads for one KRAKEN instance}

\item{inputName}{Name of input shard (FASTQ)}

\item{outputRawName}{Name of output shard (kraken raw output)}

\item{outputMatrixName}{Name of output shard (kraken count table data)}

\item{overwrite}{Force overwriting of existing files. The default is to do nothing files exist}

\item{runner}{The job manager, specifying how the command will be run (e.g. locally, or via SLURM)}

\item{bascetInstance}{A Bascet instance}
}
\value{
A job to be executed, or being executed, depending on runner settings
}
\description{
Run KRAKEN2 for each cell.
Then produce a count matrix of taxonomy IDs from the output
}
