% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{BascetRunKrakenMakeMatrix}
\alias{BascetRunKrakenMakeMatrix}
\title{TODO: should run kraken in the same run}
\usage{
BascetRunKrakenMakeMatrix(
  bascetRoot,
  useKrakenDB = "/data/henlab/kraken/standard-8",
  numLocalThreads = 1,
  inputName = "kraken_out",
  outputName = "kraken",
  runner,
  bascet_instance = bascet_instance.default
)
}
\value{
TODO
}
\description{
TODO: should run kraken in the same run
}
