% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{ChooseInformativeKMERs}
\alias{ChooseInformativeKMERs}
\title{Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range}
\usage{
ChooseInformativeKMERs(
  kmer_hist,
  num_pick = 1000,
  minfreq = 0.01,
  maxfreq = 0.1
)
}
\arguments{
\item{num_pick}{description}

\item{minfreq}{description}

\item{maxfreq}{description}

\item{fname}{description}
}
\value{
List of KMERs
}
\description{
Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range
}
