% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{ChooseInformativeKMERs}
\alias{ChooseInformativeKMERs}
\title{Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range}
\usage{
ChooseInformativeKMERs(kmerHist, minFreq = 0.005, maxFreq = 1)
}
\arguments{
\item{kmerHist}{Data.frame with KMER frequency}

\item{minFreq}{Pick KMERs having minimum frequency}

\item{maxFreq}{Pick KMERs having maxiumum frequency}
}
\value{
List of KMERs
}
\description{
Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range
}
