% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{ChooseInformativeKMERs}
\alias{ChooseInformativeKMERs}
\title{Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range}
\usage{
ChooseInformativeKMERs(kmer_hist, minfreq = 0.005, maxfreq = 1)
}
\arguments{
\item{minfreq}{description}

\item{maxfreq}{description}

\item{fname}{description}

\item{num_pick}{description}
}
\value{
List of KMERs
}
\description{
Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range
}
