% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{ChromToSpeciesCount}
\alias{ChromToSpeciesCount}
\title{Produce a count matrix on strain level}
\usage{
ChromToSpeciesCount(adata, mapSeq2strain)
}
\arguments{
\item{adata}{A Seurat object}

\item{mapSeq2strain}{A mapping from sequence to strain}
}
\value{
...
}
\description{
This function will likely be depreciated in the future, as it is a bit too specific to keep in the library
}
