% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{CloseBascet}
\alias{CloseBascet}
\title{Close a Bascet file.}
\usage{
CloseBascet(bascetFile)
}
\arguments{
\item{bascetFile}{Handle for an opened Bascet file}
}
\value{
A handle to a Bascet
}
\description{
This should always be performed to avoid memory leaks
}
