% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{CountGrangeFeatures}
\alias{CountGrangeFeatures}
\title{Obtain a feature matrix (as seurat object) given an seurat object having Fragments associated}
\usage{
CountGrangeFeatures(adata, grange_gene)
}
\arguments{
\item{adata}{Seurat object, with FeatureMatrix}

\item{grange_gene}{A grange object telling where to count}
}
\value{
A Seurat AssayObject
}
\description{
Obtain a feature matrix (as seurat object) given an seurat object having Fragments associated
}
