% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{CountSketchUMAP}
\alias{CountSketchUMAP}
\title{Run UMAP on a count sketch reduction}
\usage{
CountSketchUMAP(adata, reduction = "kmersketch", metric = "cosine", ...)
}
\arguments{
\item{adata}{A Seurat object with a count sketch reduction}

\item{reduction}{Name of the reduction to use}
}
\value{
Seurat object with UMAP computed
}
\description{
Run UMAP on a count sketch reduction
}
