% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{CreateSeuratObjectWithReduction}
\alias{CreateSeuratObjectWithReduction}
\title{Create a seurat object from e.g. count sketch reduction}
\usage{
CreateSeuratObjectWithReduction(Q, reductionName = "kmersketch", assay = "RNA")
}
\arguments{
\item{Q}{Countsketch matrix}

\item{reductionName}{Name of reduction to store Q into}

\item{assay}{Name of assay to put into reduction}
}
\value{
Seurat object holding Q as a reduction
}
\description{
Create a seurat object from e.g. count sketch reduction
}
