% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{DetectRawFileMeta}
\alias{DetectRawFileMeta}
\title{Detect metadata for raw input FASTQ files}
\usage{
DetectRawFileMeta(rawRoot, verbose = FALSE)
}
\arguments{
\item{rawRoot}{Path to folder with FASTQ files}
}
\value{
A data frame with metadata for the raw input files
}
\description{
Detect metadata for raw input FASTQ files
}
\examples{
DetectRawFileMeta("/path/to/raw_fastq", verbose = TRUE)
}
