% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{detectShardsForFile}
\alias{detectShardsForFile}
\title{Helper function: Figure out which shards belong together given root input name and extension
i.e. root/name.##.ext}
\usage{
detectShardsForFile(bascetRoot, inputName)
}
\description{
Helper function: Figure out which shards belong together given root input name and extension
i.e. root/name.##.ext
}
