% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseAMRfinder}
\alias{DownloadDatabaseAMRfinder}
\title{Download a database for AMRfinder}
\usage{
DownloadDatabaseAMRfinder(dbdir, bascetInstance = GetDefaultBascetInstance())
}
\description{
Download a database for AMRfinder
}
