% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseAriba}
\alias{DownloadDatabaseAriba}
\title{Download database for Ariba}
\usage{
DownloadDatabaseAriba(
  dbdir,
  ref = "ncbi",
  bascetInstance = GetDefaultBascetInstance()
)
}
\description{
Download database for Ariba
}
