% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseAriba}
\alias{DownloadDatabaseAriba}
\title{....}
\usage{
DownloadDatabaseAriba(
  dbdir,
  ref = c("argannot", "card", "ncbi", "megares", "plasmidfinder", "resfinder",
    "srst2_argannot", "vfdb_core", "vfdb_full", "virulencefinder"),
  bascet_instance = GetDefaultBascetInstance()
)
}
\description{
....
}
