% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{DownloadDatabaseBakta}
\alias{DownloadDatabaseBakta}
\title{....}
\usage{
DownloadDatabaseBakta(
  dbdir,
  dbtype = c("light", "full"),
  bascet_instance = GetDefaultBascetInstance()
)
}
\description{
....
}
