% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerExtractTo}
\alias{extractstreamerExtractTo}
\title{extract streamer: extract to external file}
\usage{
extractstreamerExtractTo(p, fname, outpath)
}
\arguments{
\item{p}{Streamer process}

\item{fname}{Name of file}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
0 if ok
}
\description{
extract streamer: extract to external file
}
