% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerListCellsAnyFile}
\alias{extractstreamerListCellsAnyFile}
\title{extract streamer: list all cells in a given file}
\usage{
extractstreamerListCellsAnyFile(p, fname, verbose = FALSE)
}
\arguments{
\item{p}{Streamer process}

\item{fname}{Name of file}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
List of cells
}
\description{
extract streamer: list all cells in a given file
}
