% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerLs}
\alias{extractstreamerLs}
\title{extract streamer: list all files in current Bascet zip-file}
\usage{
extractstreamerLs(p, verbose = FALSE)
}
\arguments{
\item{p}{Streamer process}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
List of files
}
\description{
extract streamer: list all files in current Bascet zip-file
}
