% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerReadNLines}
\alias{extractstreamerReadNLines}
\title{extract streamer: helper function to read N lines}
\usage{
extractstreamerReadNLines(p, n_lines, verbose = FALSE)
}
\arguments{
\item{p}{Streamer process}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
all the lines
}
\description{
extract streamer: helper function to read N lines
}
