% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerReadOneLine}
\alias{extractstreamerReadOneLine}
\title{extract streamer: helper function to read one line}
\usage{
extractstreamerReadOneLine(p, verbose = FALSE)
}
\arguments{
\item{p}{Streamer process}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
The line that was read
}
\description{
extract streamer: helper function to read one line
}
