% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{extractstreamerStart}
\alias{extractstreamerStart}
\title{extract streamer: create an instance}
\usage{
extractstreamerStart(
  fname = NULL,
  verbose = FALSE,
  bascetInstance = GetDefaultBascetInstance()
)
}
\value{
The process
}
\description{
extract streamer: create an instance
}
