% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{FragmentsToSignac}
\alias{FragmentsToSignac}
\title{From a fragments file, get a chromatin assay for Signac}
\usage{
FragmentsToSignac(fragpath)
}
\value{
TODO
}
\description{
From a fragments file, get a chromatin assay for Signac
}
