% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{FragmentsToSignac}
\alias{FragmentsToSignac}
\title{From a fragments file, get a chromatin assay for Signac  .......................... signac is dirt slow at counting! only 1 cpu used. take over the whole process}
\usage{
FragmentsToSignac(fragpath)
}
\value{
a ChromatinAssay
}
\description{
Case: tabix from command line uses 100\% cpu only. likely not designed for large queries
}
