% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetDockerImage}
\alias{getBascetDockerImage}
\title{Get a Bascet image (singularity or docker).
It will be cached in the provided directory to avoid downloading it all the time}
\usage{
getBascetDockerImage(store_at = "./", tempdir = NULL)
}
\value{
A Bascet instance
}
\description{
Get a Bascet image (singularity or docker).
It will be cached in the provided directory to avoid downloading it all the time
}
