% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{getBascetDockerImage}
\alias{getBascetDockerImage}
\title{Get and install a Bascet docker image.
It will be cached to avoid downloading it each the time the function is called}
\usage{
getBascetDockerImage(
  storeAt = getwd(),
  tempdir = NULL,
  forceInstall = FALSE,
  mapDirs = NULL,
  verbose = FALSE
)
}
\arguments{
\item{storeAt}{Directory to store the container in. Default is current directory but it is likely better to provide a single systems level directory}

\item{tempdir}{Default is to create a directory for temporary files in the current directory. Place it on a fast disk if possible}

\item{forceInstall}{Set to true to overwrite any existing Docker image}

\item{mapDirs}{Directories to map through to the container}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
A Bascet instance
}
\description{
Get and install a Bascet docker image.
It will be cached to avoid downloading it each the time the function is called
}
