% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{GetBascetTempDir}
\alias{GetBascetTempDir}
\title{Get a temp directory to use; need to be created}
\usage{
GetBascetTempDir(bascet_instance)
}
\arguments{
\item{bascet_instance}{A Bascet instance}
}
\value{
A path to a temp directory that can be created. Must be removed when done
}
\description{
Get a temp directory to use; need to be created
}
