% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{GetDefaultBascetInstance}
\alias{GetDefaultBascetInstance}
\title{Get Bascet instance from global variable}
\usage{
GetDefaultBascetInstance()
}
\value{
A Bascet instance
}
\description{
Get Bascet instance from global variable
}
