% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{GetFASTQCassembledDF}
\alias{GetFASTQCassembledDF}
\title{Get a data frame for one type of FASTQ statistics across across all cells}
\usage{
GetFASTQCassembledDF(aggr_fastq_data, section, readnum)
}
\arguments{
\item{readnum}{1 or 2, for R1 and R2}
}
\value{
TODO
}
\description{
Get a data frame for one type of FASTQ statistics across across all cells
}
