% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{is.percent.string}
\alias{is.percent.string}
\title{Given memory amount in mb, format it in a format suitable for Rust parsing}
\usage{
is.percent.string(s)
}
\description{
Check that parameter is in the form "xxx\%"
}
