% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{KmcChooseKmerFeatures}
\alias{KmcChooseKmerFeatures}
\title{Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range}
\usage{
KmcChooseKmerFeatures(fname, num_pick = 1000, minfreq = 0.01, maxfreq = 0.1)
}
\value{
TODO
}
\description{
Pick random KMERs from KMC3 database. The choice is among KMERs within a frequency range
}
