% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{KmcGetHistogram}
\alias{KmcGetHistogram}
\title{Generate a histogram from a KMC3 database}
\usage{
KmcGetHistogram(fname, plot = TRUE)
}
\value{
TODO
}
\description{
Generate a histogram from a KMC3 database
}
