% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{KneeplotPerSpecies}
\alias{KneeplotPerSpecies}
\title{Produce a kneeplot}
\usage{
KneeplotPerSpecies(adata, max_species = NULL)
}
\arguments{
\item{adata}{description}

\item{max_species}{description}
}
\value{
A ggplot object
}
\description{
Produce a kneeplot
}
