% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{KneeplotPerSpecies}
\alias{KneeplotPerSpecies}
\title{Call a MAP function for all cells}
\usage{
KneeplotPerSpecies(adata, max_species = NULL)
}
\arguments{
\item{adata}{description}

\item{max_species}{description}

\item{args}{List of arguments (key,value) to provide to the script}
}
\value{
A data.frame

TODO

A ggplot object
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{TODO override args docs
}\if{html}{\out{</div>}}
}
