% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenFindConsensusTaxonomy}
\alias{KrakenFindConsensusTaxonomy}
\title{For a Kraken count matrix, return consesus taxid for each cell as metadata}
\usage{
KrakenFindConsensusTaxonomy(mat)
}
\value{
TODO
}
\description{
For a Kraken count matrix, return consesus taxid for each cell as metadata
}
