% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenFindConsensusTaxonomy}
\alias{KrakenFindConsensusTaxonomy}
\title{For a KRAKEN2 count matrix, return consensus taxID for each cell as metadata}
\usage{
KrakenFindConsensusTaxonomy(mat)
}
\arguments{
\item{mat}{A count matrix, typically in sparse format}
}
\value{
A data.frame holding cellID and consensus taxID
}
\description{
For a KRAKEN2 count matrix, return consensus taxID for each cell as metadata
}
