% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenKneePlot}
\alias{KrakenKneePlot}
\title{Take a KRAKEN2 adata object and generate per-species kneeplots}
\usage{
KrakenKneePlot(
  adata,
  groupby = c("phylum", "class", "order", "family", "genus", "species"),
  showNumSpecies = 15,
  sortByName = FALSE
)
}
\arguments{
\item{adata}{Seurat object}

\item{groupby}{Which taxonomic level to group cells by}

\item{showNumSpecies}{Max number of species to show}

\item{sortByName}{Sort by name}
}
\value{
A ggplot object
}
\description{
Take a KRAKEN2 adata object and generate per-species kneeplots
}
