% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenKneePlot}
\alias{KrakenKneePlot}
\title{Take a KRAKEN2 adata object and generate per-species kneeplots}
\usage{
KrakenKneePlot(
  adata,
  groupby = "phylum",
  showNumSpecies = 15,
  sortByName = FALSE
)
}
\arguments{
\item{groupby}{Valid options are: Genus, phylum or species}

\item{sortByName}{Sort by name}
}
\value{
A ggplot object
}
\description{
Take a KRAKEN2 adata object and generate per-species kneeplots
}
