% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{KrakenSpeciesDistribution}
\alias{KrakenSpeciesDistribution}
\title{Using Kraken count matrix, give a "kneeplot" of species}
\usage{
KrakenSpeciesDistribution(adata, use_assay = "kraken")
}
\value{
TODO
}
\description{
Using Kraken count matrix, give a "kneeplot" of species
}
