% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{ListDatabaseAbricate}
\alias{ListDatabaseAbricate}
\title{List installed databases available for Abricate}
\usage{
ListDatabaseAbricate(dbdir, bascet_instance = GetDefaultBascetInstance())
}
\value{
List of database names
}
\description{
List installed databases available for Abricate
}
