% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_local.R
\name{LocalRunner}
\alias{LocalRunner}
\title{Create new local runner instance}
\usage{
LocalRunner(ncpu = NULL, direct = TRUE, showScript = FALSE)
}
\arguments{
\item{ncpu}{Number of CPU cores to use. By default, will attempt to detect and use all of them}

\item{direct}{Run jobs synchronously}

\item{showScript}{Show the script to run, for debugging purposes}
}
\value{
A runner instance
}
\description{
Create new local runner instance
}
