% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_local.R
\name{LocalRunner}
\alias{LocalRunner}
\title{Create new local runner instance}
\usage{
LocalRunner(maxcpu = "10", direct = FALSE, showScript = FALSE)
}
\value{
TODO
}
\description{
Create new local runner instance
}
