% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{makeOutputShardNames}
\alias{makeOutputShardNames}
\title{Helper function: Generate suitable output filenames according to shard system
i.e. root/name.##.ext}
\usage{
makeOutputShardNames(bascetRoot, outputName, ext, num_shards)
}
\description{
Helper function: Generate suitable output filenames according to shard system
i.e. root/name.##.ext
}
