% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{MapListAsDataFrame}
\alias{MapListAsDataFrame}
\title{Convenience function; alternative is to somehow implement as.data.frame}
\usage{
MapListAsDataFrame(mylist)
}
\value{
TODO
}
\description{
Convenience function; alternative is to somehow implement as.data.frame
}
