% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn_aggr.R
\name{MapListAsDataFrame}
\alias{MapListAsDataFrame}
\title{Convenience function; alternative is to somehow implement as.data.frame.}
\usage{
MapListAsDataFrame(mylist)
}
\arguments{
\item{mylist}{TODO}
}
\value{
A data.frame
}
\description{
Convenience function; alternative is to somehow implement as.data.frame.
}
