% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R
\name{MapListAsDataFrame}
\alias{MapListAsDataFrame}
\title{Convenience function for concatenating a list of data.frames into a single data.frame.
To be used to simplify the output of BascetAggregateMap}
\usage{
MapListAsDataFrame(mylist)
}
\value{
A data.frame
}
\description{
Convenience function for concatenating a list of data.frames into a single data.frame.
To be used to simplify the output of BascetAggregateMap
}
