% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorn.R, R/zorn_aggr.R
\name{MapListAsDataFrame}
\alias{MapListAsDataFrame}
\title{Convenience function for concatenating a list of data.frames into a single data.frame.
To be used to simplify the output of BascetAggregateMap}
\usage{
MapListAsDataFrame(mylist)

MapListAsDataFrame(mylist)
}
\value{
A data.frame

TODO
}
\description{
Convenience function for concatenating a list of data.frames into a single data.frame.
To be used to simplify the output of BascetAggregateMap

Convenience function; alternative is to somehow implement as.data.frame
}
