% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmatrix.R
\name{MergeBascetCountMatrix}
\alias{MergeBascetCountMatrix}
\title{Merge a list of count matrices as produced by Bascet}
\usage{
MergeBascetCountMatrix(listInput, verbose = FALSE)
}
\arguments{
\item{listInput}{List of count matrices}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
BascetCountMatrix
}
\description{
Merge a list of count matrices as produced by Bascet
}
