% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_general.R
\name{NoRunner}
\alias{NoRunner}
\title{Create new no-runner instance, used for debugging}
\usage{
NoRunner(showScript = TRUE)
}
\arguments{
\item{showScript}{For debugging: print script to run}
}
\value{
A NoRunner instance
}
\description{
Create new no-runner instance, used for debugging
}
