% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{OpenBascet}
\alias{OpenBascet}
\title{Open a Bascet, prepare it for reading individual files}
\usage{
OpenBascet(bascetRoot, bascetName, bascetInstance = GetDefaultBascetInstance())
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{bascetName}{Name of the bascet}

\item{bascetInstance}{A Bascet instance}
}
\value{
A handle to a Bascet
}
\description{
Note, the current code is based on pure R, but more efficient calls can be made
in the future. We thus advise against direct zip-file manipulation and
do not guarantee future support for this
}
