% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{OpenBascet}
\alias{OpenBascet}
\title{Open a Bascet, prepare it for reading.}
\usage{
OpenBascet(bascetRoot, bascetName)
}
\value{
TODO
}
\description{
The current code is based on pure R, but more efficient calls can be made
in the future. We thus advise against direct zip-file manipulation and
do not guarantee future support for this
}
