% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{PlotFASTQCadapterContent}
\alias{PlotFASTQCadapterContent}
\title{From aggregated FASTQC data, plot adapter content}
\usage{
PlotFASTQCadapterContent(aggrFastqData, readnum)
}
\arguments{
\item{aggrFastqData}{Aggregated FASTQ data}

\item{readnum}{1 or 2, for R1 or R2}
}
\value{
A ggplot object
}
\description{
From aggregated FASTQC data, plot adapter content
}
