% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsketch.R
\name{PlotJohnsonLindenstraussMinDim}
\alias{PlotJohnsonLindenstraussMinDim}
\title{Plot minimum number of dimensions needed to retain distance between samples}
\usage{
PlotJohnsonLindenstraussMinDim(listEps, minCells = 10, maxCells = 1e+06)
}
\arguments{
\item{listEps}{List of eps to plot for}

\item{minCells}{Plot range min cells to consider}

\item{maxCells}{Plot range max cells to consider}
}
\value{
A ggplot object
}
\description{
https://en.wikipedia.org/wiki/Johnson–Lindenstrauss_lemma
}
\details{
(1 - eps) ||u - v||^2 < ||p(u) - p(v)||^2 < (1 + eps) ||u - v||^2
}
