% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{ReadBascetCountMatrix}
\alias{ReadBascetCountMatrix}
\title{Read a count matrix as produced by Bascet (hdf5 format)}
\usage{
ReadBascetCountMatrix(fname)
}
\value{
TODO
}
\description{
Read a count matrix as produced by Bascet (hdf5 format)
}
