% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{ReadBascetCountMatrix}
\alias{ReadBascetCountMatrix}
\title{Read a count matrix as produced by Bascet (hdf5 format)}
\usage{
ReadBascetCountMatrix(fname)
}
\arguments{
\item{fname}{Full name of the hdf5 count matrix}
}
\value{
Count matrix as sparseMatrix
}
\description{
Read a count matrix as produced by Bascet (hdf5 format)
}
