% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmer.R
\name{ReadBascetCountMatrix}
\alias{ReadBascetCountMatrix}
\title{Read a count matrix as produced by Bascet (hdf5 format).
This can be output from both BascetQueryFq and BascetCountChrom}
\usage{
ReadBascetCountMatrix(bascetRoot, inputName)
}
\value{
Count matrix as sparseMatrix
}
\description{
Read a count matrix as produced by Bascet (hdf5 format).
This can be output from both BascetQueryFq and BascetCountChrom
}
