% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{ReadBascetKrakenMatrix}
\alias{ReadBascetKrakenMatrix}
\title{Read a kraken count matrix as produced by Bascet (hdf5 format)}
\usage{
ReadBascetKrakenMatrix(fname)
}
\value{
TODO
}
\description{
Read a kraken count matrix as produced by Bascet (hdf5 format)
}
