% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{ReadCellSNPmatrix}
\alias{ReadCellSNPmatrix}
\title{Read a count matrix as produced by CellSNP, but as shards}
\usage{
ReadCellSNPmatrix(bascetRoot, inputName, listCells = NULL, verbose = FALSE)
}
\value{
Count matrix as sparseMatrix
}
\description{
Read a count matrix as produced by CellSNP, but as shards
}
