% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{ReadHistogram}
\alias{ReadHistogram}
\title{Read the count histogram associated with a Bascet.
Not all Bascets have one, but it is typically produced after debarcoding}
\usage{
ReadHistogram(bascetRoot, inputName, bascet_instance = bascet_instance.default)
}
\arguments{
\item{bascet_instance}{A Bascet instance}
}
\value{
Histogram as a data.frame
}
\description{
Read the count histogram associated with a Bascet.
Not all Bascets have one, but it is typically produced after debarcoding
}
