% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascet_file.R
\name{ReadHistogram}
\alias{ReadHistogram}
\title{Read the count histogram associated with a Bascet.
Not all Bascets have one, but it is typically produced after debarcoding}
\usage{
ReadHistogram(
  bascetRoot,
  inputName,
  bascetInstance = GetDefaultBascetInstance(),
  verbose = TRUE
)
}
\arguments{
\item{bascetRoot}{The root folder where all Bascets are stored}

\item{inputName}{Name of input shard}

\item{bascetInstance}{A Bascet instance}

\item{verbose}{Print additional information, primarily to help troubleshooting}
}
\value{
Histogram as a data.frame
}
\description{
Read the count histogram associated with a Bascet.
Not all Bascets have one, but it is typically produced after debarcoding
}
