% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{SetTaxonomyNamesFeatures}
\alias{SetTaxonomyNamesFeatures}
\title{Take a KRAKEN2 count matrix where the column is the taxonomyID.
Convert to a matrix where the columns instead are the names of each taxonomy.
Unused taxonomyID columns will not be kept}
\usage{
SetTaxonomyNamesFeatures(mat, keepSpeciesOnly = TRUE)
}
\arguments{
\item{mat}{description}

\item{keepSpeciesOnly}{description}
}
\value{
A named count matrix
}
\description{
TODO: should append the name on the axis
}
\details{
TODO is this used
}
