% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscriptCancelJobIfFileExists}
\alias{shellscriptCancelJobIfFileExists}
\title{Create a piece of script to exit a job early if file exists}
\usage{
shellscriptCancelJobIfFileExists(fvar)
}
\value{
script
}
\description{
Create a piece of script to exit a job early if file exists
}
