% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shellscriptMakeFilesExpander}
\alias{shellscriptMakeFilesExpander}
\title{Generate a shell script command to produce a file of list of strings}
\usage{
shellscriptMakeFilesExpander(for_variable, list_content)
}
\arguments{
\item{for_variable}{Environment variable with array of files having the content}

\item{list_content}{Content for each file}
}
\value{
Part of a shell script
}
\description{
The name of each file will be in VARIABLE\link{TASK_ID} where TASK_ID starts from 0
}
\details{
Note that there must be one file per task, or each task deletes its own file. This will result in race conditions
}
