% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tools.R
\name{ShowFASTQCforCell}
\alias{ShowFASTQCforCell}
\title{Show the FASTQC HTML report for a cell}
\usage{
ShowFASTQCforCell(
  bascetFile,
  cellID,
  readnum = "1",
  use_browser = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{readnum}{1 or 2, for R1 and R2}

\item{use_browser}{Use operating system browser to open file}
}
\value{
TODO
}
\description{
Show the FASTQC HTML report for a cell
}
