% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_slurm.R
\name{SlurmRunner}
\alias{SlurmRunner}
\title{Create a runner that submits jobs to SLURM}
\usage{
SlurmRunner(
  settings = NULL,
  ncpu = NULL,
  partition = NULL,
  account = NULL,
  time = NULL,
  prepend = NULL,
  mem = NULL,
  direct = NULL,
  deleteScript = NULL,
  benchmark = NULL,
  verbose = NULL
)
}
\arguments{
\item{settings}{Default settings to override; can be NULL}

\item{ncpu}{Number of cores requested (SLURM -c)}

\item{partition}{Which partition to run the job on (SLURM -p)}

\item{account}{Which account to run the job on (SLURM -A)}

\item{time}{The time the job is allowed to run, e.g. "0-72:00:00" (SLURM -t)}

\item{prepend}{Something to prepend to the command. TODO seems not used. present in instance instead!!}

\item{mem}{Amount of main memory to reserve (SLURM --mem)}

\item{direct}{Run and get the result directly. FALSE implies asynchronous execution}

\item{deleteScript}{Delete job script after execution. Set to FALSE if you want to dissect it for debugging purposes}

\item{benchmark}{Enable logging of final CPU and memory usage}

\item{verbose}{Enable additional debug output}
}
\value{
A SLURM runner
}
\description{
Create a runner that submits jobs to SLURM
}
