% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{TabixGetFragmentsSeqs}
\alias{TabixGetFragmentsSeqs}
\title{Using Tabix, get list of sequences in a fragment file}
\usage{
TabixGetFragmentsSeqs(fragpath)
}
\value{
TODO
}
\description{
Using Tabix, get list of sequences in a fragment file
}
