% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refgenome.R
\name{TabixGetFragmentsSeqs}
\alias{TabixGetFragmentsSeqs}
\title{Using Tabix, get list of sequences in a fragment file}
\usage{
TabixGetFragmentsSeqs(fragpath, bascet_instance = bascet_instance.default)
}
\value{
The raw result of tabix
}
\description{
TODO too raw? wrap in bascet?
}
